% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddW.R
\name{AddW}
\alias{AddW}
\title{The Additive Weibull family}
\usage{
AddW(mu.link = "log", sigma.link = "log", nu.link = "log", tau.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}

\item{tau.link}{defines the tau.link, with "log" link as the default for the tau parameter.}
}
\description{
The Additive Weibull distribution
}
\details{
Additive Weibull distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x) = (\mu\nu x^{\nu - 1} + \sigma\tau x^{\tau - 1}) \exp({-\mu x^{\nu} - \sigma x^{\tau} }),}

for x > 0.
}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{Xie1996}{RelDists}
}
\seealso{
\link{dAddW}
}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
