% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS13.R
\name{BS13}
\alias{BS13}
\title{The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P11 Based on the fourth Tweedie)}
\usage{
BS13(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter (representing the scale \eqn{\omega}).}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter (representing the shape \eqn{\psi}).}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS13 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS13()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma} 
(where \code{mu} represents \eqn{\omega} and \code{sigma} represents \eqn{\psi}) 
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{2} \left[ \frac{x\sigma}{\mu} + \frac{\mu}{x\sigma} - 2 \right] \right) \frac{[x\sigma + \mu]}{2\sqrt{\mu x^3}}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{\mu}{\sigma} + \frac{\mu}{2\sigma^2}} and 
\eqn{Var(X) = \frac{\mu^2}{\sigma^3} + \frac{5\mu^2}{4\sigma^4}}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
set.seed(123456)
y <- rBS13(n=500, mu=5, sigma=30)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS13,
               control=gamlss.control(n.cyc=300, trace=TRUE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS13
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.1 + 1.1 * x1)      # Aprox 5
  sigma <- exp(2.2 + 2.4 * x2)   # Aprox 30
  y <- rBS13(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS13, data=dat,
               control=gamlss.control(n.cyc=500, trace=TRUE))

summary(mod2)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{dBS13}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
