% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS6.R
\name{BS6}
\alias{BS6}
\title{The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P4 Based on the mean)}
\usage{
BS6(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter.}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS6 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS6()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma} 
(where \code{mu} represents the true mean and \code{sigma} represents the shape parameter \eqn{\alpha}) 
has density given by

\eqn{f(x|\mu,\sigma) = \frac{\exp(1/\sigma^2)\sqrt{2+\sigma^2}}{4\sigma\sqrt{\pi\mu}x^{3/2}} \left[ x + \frac{2\mu}{2+\sigma^2} \right] \exp\left( -\frac{1}{2\sigma^2} \left[ \frac{\{2+\sigma^2\}x}{2\mu} + \frac{2\mu}{\{2+\sigma^2\}x} \right] \right)}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \mu} and 
\eqn{Var(X) = [\mu\sigma]^2 \left[ \frac{4+5\sigma^2}{(2+\sigma^2)^2} \right]}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
set.seed(1234)
y <- rBS6(n=50, mu=1, sigma=0.1)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS6)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS6
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.5 - 3 * x1)        # Aprox 1
  sigma <- exp(0.5 - 3.5 * x2)   # Aprox 0.1
  y <- rBS6(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

dat <- gendat(n=100)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS6, data=dat)

summary(mod2)
}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{dBS6}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
