% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS8.R
\name{BS8}
\alias{BS8}
\title{The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P6 Based on the variance 2)}
\usage{
BS8(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter (representing the variance).}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter (representing the shape).}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS8 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS8()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma} 
(where \code{mu} represents the true variance \eqn{\sigma^2} and \code{sigma} represents the shape parameter \eqn{\alpha}) 
has density given by

\eqn{f(x|\mu,\sigma) = 
\frac{\sqrt{\mu}}
     {2\sqrt{2\pi\sigma}}
\left[
\left\{
\frac{1}{2x}
\sqrt{\frac{5\sigma}{\mu(\mu-1)}}
\right\}^{1/2}
+
\left\{
\frac{1}{2x}
\sqrt{\frac{5\sigma}{\mu(\mu-1)}}
\right\}^{3/2}
\right]
\exp\left(
-\frac{5}{8(\mu-1)}
\left[
\frac{2x\sqrt{\mu(\mu-1)}}{\sqrt{5\sigma}}
+
\frac{\sqrt{5\sigma}}
     {2+\sqrt{\mu(\mu-1)}}
-2
\right]
\right)
}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{[2\mu+3]\sqrt{\sigma}}{\sqrt{20\mu(\mu-1)}}} and 
\eqn{Var(X) = \sigma}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma

set.seed(12345)
y <- rBS8(n=100, mu=2, sigma=10)

# 1. Extract the built-in "logshiftto1" link structure
logshift_link <- make.link.gamlss("logshiftto1")

# 2. Assign its components to the 'own' functions gamlss searches for
own.linkfun  <- logshift_link$linkfun
own.linkinv  <- logshift_link$linkinv
own.mu.eta   <- logshift_link$mu.eta
own.valideta <- logshift_link$valideta

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS8(mu.link = "own"),
               control=gamlss.control(n.cyc=100))

mod99 <- gamlss(y~1, sigma.fo=~1, family=BS8,
               control=gamlss.control(n.cyc=100))

# Extracting the fitted values for mu and sigma
# using the inverse link function
own.linkinv(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

exp(coef(mod99, what="mu"))
exp(coef(mod99, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS8
\dontrun{
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.6 - 1.4 * x1)      # Aprox 2.45
  sigma <- exp(1.6 + 1.5 * x2)   # Aprox 10
  y <- rBS8(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS8(mu.link = "own"), data=dat,
               control=gamlss.control(n.cyc=100))

summary(mod2)
}
}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{dBS8}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
