% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dAddW.R
\name{dAddW}
\alias{dAddW}
\alias{pAddW}
\alias{qAddW}
\alias{rAddW}
\alias{hAddW}
\title{The Additive Weibull distribution}
\usage{
dAddW(x, mu, sigma, nu, tau, log = FALSE)

pAddW(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qAddW(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rAddW(n, mu, sigma, nu, tau)

hAddW(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{shape parameter.}

\item{tau}{shape parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dAddW} gives the density, \code{pAddW} gives the distribution 
function, \code{qAddW} gives the quantile function, \code{rAddW}
generates random deviates and \code{hAddW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Additive Weibull distribution
with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
Additive Weibull Distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x) = (\mu\nu x^{\nu - 1} + \sigma\tau x^{\tau - 1}) \exp({-\mu x^{\nu} - \sigma x^{\tau} }),}

for x > 0.
}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{Xie1996}{RelDists}
}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
