% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dAPIE.R
\name{dAPIE}
\alias{dAPIE}
\alias{pAPIE}
\alias{qAPIE}
\alias{rAPIE}
\alias{hAPIE}
\title{The Alpha Power Inverted Exponential distribution}
\usage{
dAPIE(x, mu, sigma, log = FALSE)

pAPIE(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)

qAPIE(p, mu, sigma, lower.tail = TRUE, log.p = FALSE)

rAPIE(n, mu, sigma)

hAPIE(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dAPIE} gives the density, \code{pAPIE} gives the distribution 
function, \code{qAPIE} gives the quantile function, \code{rAPIE}
generates random deviates and \code{hAPIE} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Alpha Power Inverted Exponential distribution
with parameters \code{mu} and \code{sigma}.
}
\details{
The Alpha Power Inverted Exponential distribution \code{mu} and 
\code{sigma}  has density given by

\eqn{f(x) = \frac{log(\mu)}{\mu - 1} \frac{\sigma}{x^{2}} \exp(-\frac{\sigma}{x}) 
\mu^{\exp(-\frac{\sigma}{x})},}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\mu \neq 1}.

\eqn{f(x) = \frac{\sigma}{x^{2}} \exp(-\frac{\sigma}{x}),}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\mu = 1}.
}
\examples{
 
## The probability density function
curve(dAPIE(x, mu=3, sigma=0.5), from=0.001, to=10,
      ylim=c(0, 0.8), col="red", ylab="f(x)", las=1)

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pAPIE(x,  mu=3, sigma=0.5),
      from=0.0001, to=25, col="red", las=1, ylab="F(x)")
curve(pAPIE(x,  mu=3, sigma=0.5, lower.tail=FALSE),
      from=0.0001, to=25, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qAPIE(p,  mu=3, sigma=0.5), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pAPIE(x,  mu=3, sigma=0.5),
      from=0.001, add=TRUE, col="red")

## The random function
hist(rAPIE(n=1000,  mu=3, sigma=0.5), freq=FALSE,breaks=50,
     xlab="x", ylim=c(0, 0.033), las=1, main="", xlim=c(0, 200))
curve(dAPIE(x,  mu=3, sigma=0.5),
      from=0.001, to=500, add=TRUE, col="red")

## The Hazard function
curve(hAPIE(x,  mu=3, sigma=0.5), from=0.001, to=15,
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{ceren2018}{RelDists}
}
\author{
Santiago Restrepo Alvarez
}
