% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBS11.R
\name{dBS11}
\alias{dBS11}
\alias{pBS11}
\alias{qBS11}
\alias{rBS11}
\alias{hBS11}
\title{The Birnbaum-Saunders distribution - Santos-Neto et al. (2012) (P9 Based on the second Tweedie)}
\usage{
dBS11(x, mu = 1, sigma = 0.5, log = FALSE)

pBS11(q, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

qBS11(p, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

rBS11(n, mu = 1, sigma = 0.5)

hBS11(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter representing \eqn{\beta} (\code{mu > 0}).}

\item{sigma}{parameter representing \eqn{\omega} (\code{sigma > 0}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dBS11} gives the density, \code{pBS11} gives the distribution 
function, \code{qBS11} gives the quantile function, \code{rBS11}
generates random deviates and \code{hBS11} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the 
Birnbaum-Saunders distribution with 
parameters \code{mu} and \code{sigma}.
}
\details{
The Birnbaum-Saunders with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{2\mu} \left[ \frac{x}{\mu} + \frac{\mu}{x} - 2 \right] \right) \frac{[x + \mu]\sqrt{\sigma}}{2\mu\sqrt{x^3}}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \mu + \frac{\mu^2}{2\sigma}} and 
\eqn{Var(X) = \frac{\mu^3}{\sigma} + \frac{5\mu^4}{4\sigma^2}}.
}
\examples{
# Example 1
# Plotting the mass function for different parameter values
curve(dBS11(x, mu=1, sigma=12), 
      from=0.001, to=2.5,
      col="royalblue1", lwd=2, 
      main="Density function",
      xlab="x", ylab="f(x)")
curve(dBS11(x, mu=1, sigma=0.5),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("topright", legend=c("mu=1, sigma=12", 
                            "mu=1, sigma=0.5"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.6)

# Example 2
# Checking if the cumulative curves converge to 1
curve(pBS11(x, mu=1, sigma=12), 
      from=0.00001, to=8,
      ylim=c(0, 1), 
      col="royalblue1", lwd=2, 
      main="Cumulative Distribution Function",
      xlab="x", ylab="F(x)")
curve(pBS11(x, mu=1, sigma=0.5),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("bottomright", legend=c("mu=1, sigma=12", 
                               "mu=1, sigma=0.5"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.5)

# Example 3
# The quantile function
p <- seq(from=0, to=0.999, length.out=100)
plot(x=qBS11(p, mu=1, sigma=12), y=p, xlab="Quantile",
     las=1, ylab="Probability", main="Quantile function ")
curve(pBS11(x, mu=1, sigma=12), 
      from=0, add=TRUE, col="tomato", lwd=2.5)

# Example 4
# The random function
x <- rBS11(n=10000, mu=1, sigma=12)
hist(x, freq=FALSE)
curve(dBS11(x, mu=1, sigma=12),  
      add=TRUE, col="tomato", lwd=2)

# Example 5
# The Hazard function
curve(hBS11(x, mu=1, sigma=12), from=0.001, to=4,
      col="tomato", ylab="Hazard function", las=1)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{BS11}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
