% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBS13.R
\name{dBS13}
\alias{dBS13}
\alias{pBS13}
\alias{qBS13}
\alias{rBS13}
\alias{hBS13}
\title{The Birnbaum-Saunders distribution - Santos-Neto et al. (2012) (P11 Based on the fourth Tweedie)}
\usage{
dBS13(x, mu = 1, sigma = 0.5, log = FALSE)

pBS13(q, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

qBS13(p, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

rBS13(n, mu = 1, sigma = 0.5)

hBS13(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter representing \eqn{\omega} (\code{mu > 0}).}

\item{sigma}{parameter representing \eqn{\psi} (\code{sigma > 0}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dBS13} gives the density, \code{pBS13} gives the distribution 
function, \code{qBS13} gives the quantile function, \code{rBS13}
generates random deviates and \code{hBS13} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the 
Birnbaum-Saunders distribution with 
parameters \code{mu} and \code{sigma}.
}
\details{
The Birnbaum-Saunders with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{2} \left[ \frac{x\sigma}{\mu} + \frac{\mu}{x\sigma} - 2 \right] \right) \frac{[x\sigma + \mu]}{2\sqrt{\mu x^3}}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{\mu}{\sigma} + \frac{\mu}{2\sigma^2}} and 
\eqn{Var(X) = \frac{\mu^2}{\sigma^3} + \frac{5\mu^2}{4\sigma^4}}.
}
\examples{
# Example 1
# Plotting the mass function for different parameter values
curve(dBS13(x, mu=5, sigma=30), 
      from=0.001, to=0.8,
      col="royalblue1", lwd=2, 
      main="Density function",
      xlab="x", ylab="f(x)")
curve(dBS13(x, mu=5, sigma=10),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("topright", legend=c("mu=5, sigma=30", 
                            "mu=5, sigma=10"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.6)

# Example 2
# Checking if the cumulative curves converge to 1
curve(pBS13(x, mu=5, sigma=30), 
      from=0.00001, to=2,
      ylim=c(0, 1), 
      col="royalblue1", lwd=2, 
      main="Cumulative Distribution Function",
      xlab="x", ylab="F(x)")
curve(pBS13(x, mu=5, sigma=10),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("bottomright", legend=c("mu=5, sigma=30", 
                               "mu=5, sigma=10"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.5)

# Example 3
# The quantile function
p <- seq(from=0, to=0.999, length.out=100)
plot(x=qBS13(p, mu=5, sigma=30), y=p, xlab="Quantile",
     las=1, ylab="Probability", main="Quantile function ")
curve(pBS13(x, mu=5, sigma=30), 
      from=0, add=TRUE, col="tomato", lwd=2.5)

# Example 4
# The random function
x <- rBS13(n=10000, mu=5, sigma=30)
hist(x, freq=FALSE)
curve(dBS13(x, mu=5, sigma=30),  
      add=TRUE, col="tomato", lwd=2)

# Example 5
# The Hazard function
curve(hBS13(x, mu=5, sigma=30), from=0.001, to=1,
      col="tomato", ylab="Hazard function", las=1)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{BS13}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
