% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBW.R
\name{dBW}
\alias{dBW}
\alias{pBW}
\alias{qBW}
\alias{rBW}
\alias{hBW}
\title{The Beta Weibull distribution}
\usage{
dBW(x, mu, sigma, nu, tau, log = FALSE)

pBW(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qBW(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rBW(n, mu, sigma, nu, tau)

hBW(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{tau}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dBW} gives the density, \code{pBW} gives the distribution 
function, \code{qBW} gives the quantile function, \code{rBW}
generates random deviates and \code{hBW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Beta Weibull distribution 
with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
The Beta Weibull Distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x)= \frac{1}{\beta \left( \nu, \tau \right)} \mu \sigma x^{\sigma - 1} (1 - \exp(-\mu x^\sigma))^{\nu - 1} \exp(-\mu \tau x^\sigma),}

for x > 0.
}
\examples{
 
## The probability density function 
curve(dBW(x, mu = (1/4), sigma =1, nu=1, tau=2), from = 0, to = 10, 
      col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pBW(x, mu = (1/4), sigma =1, nu=1, tau=2), from = 0, to = 6, 
ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pBW(x, mu = (1/4), sigma =1, nu=1, tau=2, lower.tail = FALSE), 
from = 0, to = 6, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qBW(p = p, mu = (1/4), sigma =1, nu=1, tau=2), y = p, 
xlab = "Quantile", las = 1, ylab = "Probability")
curve(pBW(x, mu = (1/4), sigma =1, nu=1, tau=2), from = 0, add = TRUE, 
col = "red")

## The random function
hist(rBW(1000, mu = (1/4), sigma =1, nu=1, tau=2), freq = FALSE, xlab = "x", 
ylim = c(0, 0.5), las = 1, main = "")
curve(dBW(x, mu = (1/4), sigma =1, nu=1, tau=2),  from = 0, add = TRUE, 
col = "red", ylim = c(0, 0.5))

## The Hazard function(
par(mfrow=c(1,1))
curve(hBW(x, mu = (1/4), sigma =1, nu=1, tau=2), from = 0, to = 2, 
ylim = c(0, 1), col = "red", ylab = "The hazard function", las = 1)

}
