% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEC.R
\name{dEC}
\alias{dEC}
\alias{pEC}
\alias{qEC}
\alias{rEC}
\alias{hEC}
\title{The Exponentiated Chen distribution}
\usage{
dEC(x, mu, sigma, nu, log = FALSE)

pEC(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qEC(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rEC(n, mu, sigma, nu)

hEC(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEC} gives the density, \code{pEC} gives the distribution 
function, \code{qEC} gives the quantile function, \code{rEC}
generates random deviates and \code{hEC} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for The Exponentiated Chen distribution 
with parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The Exponentiated Chen distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\mu \sigma \nu x^{\sigma - 1} \exp({x ^\sigma}) \exp\{{\nu (1 -\exp({x^\sigma})}\} 
\{1 - \exp({\nu (1 -\exp({x^\sigma}))})\}^{\mu-1},}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\nu> 0}.
}
\examples{
## The probability density function
curve(dEC(x, mu=35, sigma=0.25, nu=2), from=0.1, to=10,
      ylim=c(0, 1.5), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pEC(x, mu=35, sigma=0.25, nu=2),
      from=0.1, to=10,  col="red", las=1, ylab="F(x)")
curve(pEC(x, mu=35, sigma=0.25, nu=2, lower.tail=FALSE),
      from=0.1, to=10, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qEC(p, mu=35, sigma=0.25, nu=2), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pEC(x,  mu=35, sigma=0.25, nu=2), from=0.1, add=TRUE, col="red")

## The random function
hist(rEC(n=10000, mu=35, sigma=0.25, nu=2), freq=FALSE,
     xlab="x", las=1, main="")
curve(dEC(x, mu=35, sigma=0.25, nu=2), from=0.1, add=TRUE, col="red")

## The Hazard function
curve(hEC(x, mu=35, sigma=0.25, nu=2), from=0.1, to=30, ylim=c(0, 1),
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{sanku2017}{RelDists}
}
\author{
Santiago Toro
}
