% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEEB.R
\name{dEEB}
\alias{dEEB}
\alias{pEEB}
\alias{qEEB}
\alias{rEEB}
\alias{hEEB}
\title{The Exponentiated Exponential Binomial distribution}
\usage{
dEEB(x, mu, sigma, nu, size, log = FALSE)

pEEB(q, mu, sigma, nu, size, lower.tail = TRUE, log.p = FALSE)

qEEB(p, mu, sigma, nu, size, lower.tail = TRUE, log.p = FALSE)

rEEB(n, mu, sigma, nu, size)

hEEB(x, mu, sigma, nu, size)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scale parameter.}

\item{sigma, nu}{shape parameters.}

\item{size}{number of trials (one or more).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEEB} gives the density, \code{pEEB} gives the distribution 
function, \code{qEEB} gives the quantile function, \code{rEEB}
generates random deviates and \code{hEEB} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the odd Weibull distribution with
parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The exponentiated exponential binomial distribution with parameters \code{size =}\eqn{m}, 
\code{mu}, \code{sigma} and \code{nu} has density given by

\eqn{f(x) = \sigma\nu\mu m \exp(-\sigma x) (1 - \exp(-\sigma x) )^(\nu-1) (1 - \mu(1 - \exp(-\sigma x))^\nu)^(m-1) ,}

for \eqn{x > 0}, \eqn{0 < \mu < 1}, \eqn{\sigma > 0} and \eqn{\nu >0}.
}
\examples{
 
## The probability density function
par(mfrow = c(1, 2))
curve(dEEB(x, mu=1, sigma=1, nu=0.5, size=10), from=0, to=1.5, ylim=c(0, 1.5), 
      col="red", las=1, ylab="The probability density function")
curve(dEEB(x, mu=1, sigma=1, nu=5, size=10), from=0, to=1.5, ylim=c(0, 3), 
      col="red", las=1, ylab="The probability density function")
      
## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pEEB(x, mu=1, sigma=1, nu=5, size=10), from=0, to=4, ylim=c(0, 1), 
      col="red", las = 1, ylab = "The cumulative distribution function")
curve(pEEB(x, mu=1, sigma=1, nu=5, size=10, lower.tail=FALSE), from=0, to=4,  
      ylim=c(0, 1), col="red", las=1, ylab="The Reliability function")

## The quantile function
p <- seq(from=0, to=0.998, length.out=100)
plot(x=qEEB(p, mu=1, sigma=1, nu=5, size=10), y=p, xlab="Quantile", 
    las=1, ylab="Probability")
curve(pEEB(x, mu=1, sigma=1, nu=5, size=10), from=0, add=TRUE, col="red")

## The random function
hist(rEEB(n=10000, mu=1, sigma=1, nu=5, size=10), freq=FALSE, 
     ylim = c(0,3),xlab="x", las=1, main="")
curve(dEEB(x, mu=1, sigma=1, nu=5, size=10), from=0, ylim=c(0, 3), 
      add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1,2))
curve(hEEB(x, mu=1, sigma=1, nu=0.5, size=10), from=0, to=4, ylim=c(0, 3), 
      col="red", ylab="The hazard function", las=1)
curve(hEEB(x, mu=1, sigma=1, nu=5, size=10), from=0, to=6, ylim=c(0, 3), 
      col="red", ylab="The hazard function", las=1)
      
}
\references{
\insertRef{Bakouch2012}{RelDists}
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
