% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{den_exw}
\alias{den_exw}
\title{Auxiliar den_exw function for the Ex-Wald distribution}
\usage{
den_exw(r, m, a, t)
}
\arguments{
\item{r}{value for r.}

\item{m}{value for m.}

\item{a}{value for a.}

\item{t}{value for t.}
}
\value{
returns a vector with probabilities.
}
\description{
This function emulates the dExWALD function.
}
\keyword{internal}
