% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEpsilon.R
\name{dEpsilon}
\alias{dEpsilon}
\alias{pEpsilon}
\alias{qEpsilon}
\alias{rEpsilon}
\alias{hEpsilon}
\title{The Epsilon distribution}
\usage{
dEpsilon(x, mu, sigma, log = FALSE)

pEpsilon(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)

qEpsilon(p, mu, sigma, lower.tail = TRUE, log.p = FALSE)

rEpsilon(n, mu, sigma)

hEpsilon(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEpsilon} gives the density, \code{pEpsilon} gives the distribution 
function, \code{qEpsilon} gives the quantile function, \code{rEpsilon}
generates random deviates and \code{hEpsilon} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Epsilon distribution 
with parameters \code{mu} and \code{sigma}.
}
\details{
The Epsilon Distribution with parameters \code{mu} and 
\code{sigma} has density given by

\eqn{f(x)= \mu \frac{\sigma^2}{\sigma^2-x^2}\varepsilon_{-\mu,\sigma}\left(x\right)}

where

\eqn{\varepsilon_{-\mu,\sigma}(x) = \left(\frac{x+\sigma}{\sigma-x}\right)^{-\mu\dfrac{\sigma}{2}}}

for \eqn{0 < x < \sigma}, \eqn{\mu > 0}  and \eqn{\sigma > 0}.
}
\examples{
 
## The probability density function
par(mfrow=c(1,1))
curve(dEpsilon(x, mu=0.7, sigma=5), from=0, to=4.999,
      ylim=c(0, 0.7), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pEpsilon(x, mu=0.7, sigma=5),
      from=0, to=4.999,  col="red", las=1, ylab="F(x)")
curve(pEpsilon(x, mu=0.7, sigma=5, lower.tail=FALSE),
      from=0, to=4.999, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qEpsilon(p, mu=0.7, sigma=5), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pEpsilon(x, mu=0.7, sigma=10), from=0, add=TRUE, col="red")

## The random function
hist(rEpsilon(n=10000, mu=0.7, sigma=5), freq=FALSE,
     xlab="x", las=1, main="")
curve(dEpsilon(x, mu=0.7, sigma=10), from=0, add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1,1))
curve(hEpsilon(x, mu=0.7, sigma=5), from=0, to=4.999, ylim=c(0, 15),
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{dombi2018epsilon}{RelDists}
}
\author{
Daniel Betancur
}
