% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EW.R
\name{dEW}
\alias{dEW}
\alias{pEW}
\alias{qEW}
\alias{rEW}
\alias{hEW}
\title{The Exponentiated Weibull Distribution}
\usage{
dEW(x, mu, sigma, nu, log = FALSE)

pEW(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qEW(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rEW(n, mu, sigma, nu)

hEW(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scale parameter.}

\item{sigma, nu}{shape parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

#' @details 
The Exponentiated Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\nu \mu \sigma x^{\sigma-1} \exp(-\mu x^\sigma) (1-\exp(-\mu x^\sigma))^{\nu-1},}

for x > 0.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEW} gives the density, \code{pEW} gives the distribution 
function, \code{qEW} gives the quantile function, \code{rEW}
generates random deviates and \code{hEW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation  and hazard function for the exponentiated weibull distribution with
parameters \code{mu}, \code{sigma} and \code{nu}.
}
\examples{
 
## The probability density function
curve(dEW(x, mu=2, sigma=1.5, nu=0.5), from=0, to=2,
      ylim=c(0, 2.5), col="red", las=1, ylab="The probability density function") 

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pEW(x, mu=2, sigma=1.5, nu=0.5), from=0, to=2,  col="red",
      las=1, ylab="The cumulative distribution function")
curve(pEW(x, mu=2, sigma=1.5, nu=0.5, lower.tail=FALSE), from=0,
      to=2, col="red", las=1, ylab="The Reliability function")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qEW(p, mu=2, sigma=1.5, nu=0.5), y=p, xlab="Quantile", las=1, ylab="Probability")
curve(pEW(x, mu=2, sigma=1.5, nu=0.5),  from=0, add=TRUE, col="red")

## The random function
hist(rEW(n=10000, mu=2, sigma=1.5, nu=0.5), freq=FALSE, xlab="x", las=1, main="")
curve(dEW(x, mu=2, sigma=1.5, nu=0.5),  from=0, add=TRUE, col="red") 

## The Hazard function
curve(hEW(x, mu=2,sigma=1.5, nu=0.5), from=0, to=2, ylim=c(0, 7), col="red",
      ylab="The Hazard function")

}
