% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEWP.R
\name{dEWP}
\alias{dEWP}
\alias{pEWP}
\alias{qEWP}
\alias{rEWP}
\alias{hEWP}
\title{The Exponentieted Weibull-Poisson distribution}
\usage{
dEWP(x, mu, sigma, nu, tau, log = FALSE)

pEWP(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qEWP(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rEWP(n, mu, sigma, nu, tau)

hEWP(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scale parameter.}

\item{sigma, nu, tau}{shape parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dEWP} gives the density, \code{pEWP} gives the distribution 
function, \code{qEWP} gives the quantile function, \code{rEWP}
generates random deviates and \code{hEWP} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the odd Weibull distribution with
parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
The exponentieted Weibull-Poisson distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x)=\sigma \nu \tau \mu^\nu x^(\nu - 1) \exp(-(\mu x)^\nu) (1 - \exp(-(\mu x)^\nu))^(\sigma - 1)
           \exp(\tau(1 - \exp(-(\mu x)^\nu))^\sigma),}
           
for \eqn{x > 0}, \eqn{\mu > 0}, \eqn{\sigma > 0}, \eqn{\nu >0} and \eqn{\nu >0}.
}
\examples{
 
## The probability density function
par(mfrow = c(1, 2))
curve(dEWP(x, mu=1, sigma=0.5, nu=1, tau=5), from=0, to=4, ylim=c(0, 0.5), 
      col="red", las=1, ylab="The probability density function")
curve(dEWP(x, mu=1, sigma=0.5, nu=3, tau=0.5), from=0, to=2, ylim=c(0, 1), 
      col="red", las=1, ylab="The probability density function")
      
## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pEWP(x, mu=1, sigma=0.5, nu=1, tau=5), from=0, to=4, ylim=c(0, 1), 
      col="red", las = 1, ylab = "The cumulative distribution function")
curve(pEWP(x, mu=1, sigma=0.5, nu=1, tau=5, lower.tail=FALSE), from=0, to=10,  
      ylim=c(0, 1), col="red", las=1, ylab="The Reliability function")

}
\references{
\insertRef{Mahmoudi2013}{RelDists}
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
