% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dExWALD.R
\name{dExWALD}
\alias{dExWALD}
\alias{pExWALD}
\alias{qExWALD}
\alias{rExWALD}
\title{The Ex-Wald distribution}
\usage{
dExWALD(x, mu = 1.5, sigma = 1.5, nu = 2, log = FALSE)

pExWALD(q, mu = 1.5, sigma = 1.5, nu = 2, lower.tail = TRUE, log.p = FALSE)

qExWALD(p, mu = 1.5, sigma = 1.5, nu = 2)

rExWALD(n, mu = 1.5, sigma = 1.5, nu = 2)
}
\arguments{
\item{x, q}{vector of (non-negative integer) quantiles.}

\item{mu}{vector of the mu parameter.}

\item{sigma}{vector of the sigma parameter.}

\item{nu}{vector of the nu parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X <= x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values to return.}
}
\value{
\code{dExWALD} gives the density, \code{pExWALD} gives the distribution
function, \code{qExWALD} gives the quantile function, \code{rExWALD}
generates random deviates.
}
\description{
These functions define the density, distribution function, quantile
function and random generation for the Ex-Wald distribution
with parameter \eqn{\mu}, \eqn{\sigma} and \eqn{\nu}.
}
\details{
The Wald distribution with parameters \eqn{\mu}, \eqn{\sigma} 
and \eqn{\nu} has density given by

\eqn{f(x |\mu, \sigma)= x+1}
}
\examples{
## The probability density function

curve(dExWALD(x, mu=1, sigma=4, nu=1), from=0, to=10, 
      ylim=c(0, 0.3), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
curve(pExWALD(x, mu=1, sigma=4, nu=1), from=0, to=10, 
      col="red", las=1, ylab="F(x)")

curve(pExWALD(x, mu=1, sigma=4, nu=1, lower.tail=FALSE), 
      from=0, to=10, col="red", las=1, ylab="R(x)")

## The quantile function
p <- seq(from=0.01, to=0.99, length.out=100)
plot(x=qExWALD(p, mu=1.5, sigma=1.5, nu=2), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pExWALD(x, mu=1.5, sigma=1.5, nu=2), from=0, add=TRUE, col="red")

## The random function
x <- rExWALD(n=1000, mu=1.5, sigma=1.5, nu=2)
hist(x, freq=FALSE, xlab="x", 
     ylim=c(0, 0.35), las=1, main="")
curve(dExWALD(x, mu=1.5, sigma=1.5, nu=2), 
      from=0, to=20, add=TRUE, col="red")


}
\references{
Heathcote, A. (2004). Fitting Wald and ex-Wald distributions to 
response time data: An example using functions for the S-PLUS package. 
Behavior Research Methods, Instruments, & Computers, 36, 678-694.
}
\seealso{
\link{ExWALD}.
}
\author{
Freddy Hernandez, \email{fhernanb@unal.edu.co}
}
