% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dFPEGG.R
\name{dFPEGG}
\alias{dFPEGG}
\alias{pFPEGG}
\alias{qFPEGG}
\alias{rFPEGG}
\alias{hFPEGG}
\title{The Four-Parameter Exponentiated Generalized Gamma distribution}
\usage{
dFPEGG(x, mu, sigma, nu, tau, log = FALSE)

pFPEGG(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qFPEGG(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rFPEGG(n, mu, sigma, nu, tau)

hFPEGG(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{tau}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dFPEGG} gives the density, \code{pFPEGG} gives the distribution 
function, \code{qFPEGG} gives the quantile function, \code{rFPEGG}
generates random deviates and \code{hFPEGG} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Four-Parameter Exponentiated Generalized Gamma distribution
with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
Four-Parameter Exponentiated Generalized Gamma distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x) = \frac {\nu \sigma}{\mu \Gamma(\tau)} (\frac {x}{\mu})^{\sigma \tau -1} exp\{(-{\frac {x}{\mu})^{\sigma}})\} 
\{ \gamma~_1~ (\tau, (\frac {x}{\mu})^{\sigma}) \}^{\nu -1} ,}

for x > 0.
}
\examples{
## The probability density function
curve(dFPEGG(x, mu=0.1, sigma=0.8, nu=10, tau=1.5), from=0.000001, to=1.5, ylim=c(0, 2.5),
      col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pFPEGG(x, mu=0.1, sigma=0.8, nu=10, tau=1.5),
      from=0.000001, to=1.5, col="red", las=1, ylab="F(x)")
curve(pFPEGG(x, mu=0.1, sigma=0.8, nu=10, tau=1.5, lower.tail=FALSE),
      from=0.000001, to=1.5, col="red", las=1, ylab="R(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qFPEGG(p, mu=0.1, sigma=0.8, nu=10, tau=1.5), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pFPEGG(x, mu=0.1, sigma=0.8, nu=10, tau=1.5), 
      from=0.00001, add=TRUE, col="red")

## The random function
hist(rFPEGG(n=100, mu=0.1, sigma=0.8, nu=10, tau=1.5), freq=FALSE,
     xlab="x", las=1, main="")
curve(dFPEGG(x, mu=0.1, sigma=0.8, nu=10, tau=1.5),
      from=0.0001, to=2, add=TRUE, col="red")

## The Hazard function
curve(hFPEGG(x,  mu=0.1, sigma=0.8, nu=10, tau=1.5), from=0.0001, to=1.5,
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{cordeiro2011}{RelDists}
}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
