% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dFPGG.R
\name{dFPGG}
\alias{dFPGG}
\alias{pFPGG}
\alias{qFPGG}
\alias{rFPGG}
\alias{hFPGG}
\title{The Four-Parameter Generalized Gamma distribution}
\usage{
dFPGG(x, mu, sigma, nu, tau, log = FALSE)

pFPGG(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qFPGG(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rFPGG(n, mu, sigma, nu, tau)

hFPGG(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{tau}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dFPGG} gives the density, \code{pFPGG} gives the distribution 
function, \code{qFPGG} gives the quantile function, \code{rFPGG}
generates random deviates and \code{hFPGG} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Four-Parameter Generalized Gamma distribution
with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
Four-Parameter Generalized Gamma distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x) = \frac {\sigma \mu^{\tau \sigma}}{\tau (\tau)} (x - \nu)^{\tau \sigma -1} \exp-({\frac {(x-\nu)} {\mu} })^{\sigma},}

for x >= 0.
}
\examples{
 
## The probability density function
curve(dFPGG(x, mu=1, sigma=1, nu=0, tau=4), from=0.1, to=15,
      col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pFPGG(x, mu=1, sigma=1, nu=0, tau=4),
      from=0, to=4, col="red", las=1, ylab="F(x)")
curve(pFPGG(x,  mu=1, sigma=1, nu=0, tau=4, lower.tail=FALSE),
      from=0, to=4, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qFPGG(p, mu=1, sigma=1, nu=0, tau=4), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pFPGG(x,  mu=1, sigma=1, nu=0, tau=4), 
      from=0, add=TRUE, col="red")

hist(rFPGG(n=10000,  mu=1, sigma=1, nu=0, tau=4), freq=FALSE,
     xlab="x", las=1, main="")
curve(dFPGG(x,  mu=1, sigma=1, nu=0, tau=4),
      from=0, to=15, add=TRUE, col="red")

## The Hazard function
curve(hFPGG(x,  mu=1, sigma=1, nu=0, tau=4), from=0, to=30,
      col="red", ylab="Hazard function", las=1)

}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
