% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGPW.R
\name{dGPW}
\alias{dGPW}
\alias{pGPW}
\alias{qGPW}
\alias{rGPW}
\alias{hGPW}
\title{The Generalized power Weibull Distribution}
\usage{
dGPW(x, mu, sigma, nu, log = FALSE)

pGPW(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qGPW(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rGPW(n, mu, sigma, nu)

hGPW(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameters.}

\item{nu}{parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dGPW} gives the density, \code{pGPW} gives the distribution 
function, \code{qGPW} gives the quantile function, \code{rGPW}
generates random deviates and \code{hGPW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Generalized power Weibull Distribution with
parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The Generalized power Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\mu\sigma\nu^{-1} x^{\sigma-1}(1+\mu x^\sigma)^{\frac 1{\nu}-1}\;exp{[1-(1+\mu x^\sigma)^{\frac 1{\nu}}]}}

for x > 0.
}
\examples{
 

## The probability density function
curve(dGPW(x, mu=0.2, sigma=2.0, nu=0.3), 
     xlim=c(0,2), ylim=c(0, 2.5), col="red", las=1, ylab="f(x)")


## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pGPW(x, mu=0.2, sigma=2.0, nu=0.3), 
     xlim=c(0,2), col="red", las=1, ylab="F(x)")
curve(pGPW(x, mu=2, sigma=1.5, nu=0.5, lower.tail=FALSE),
     from=0, to=2, col="red", las=1, ylab="R(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qGPW(p, mu=0.2, sigma=2.0, nu=0.3), y=p, xlab="Quantile",
    las=1, ylab="Probability")
curve(pGPW(x, mu=0.2, sigma=2.0, nu=0.3), xlim=c(0,2), add=TRUE, col="red")

## The random function
hist(rGPW(n=10000, mu=0.2, sigma=2.0, nu=0.3), freq=FALSE,
    xlab="x", las=1, main="")
curve(dGPW(x, mu=0.2, sigma=2.0, nu=0.3), xlim=c(0,2), add=TRUE, col="red")

## The Hazard function
curve(hGPW(x, mu=0.2, sigma=2.0, nu=0.3), xlim=c(0,2), ylim=c(0, 7),
     col="red", ylab="Hazard function", las=1)


}
