% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGW.R
\name{dGW}
\alias{dGW}
\alias{pGW}
\alias{qGW}
\alias{rGW}
\alias{hGW}
\title{The Gamma Weibull distribution}
\usage{
dGW(x, mu, sigma, nu, log = FALSE)

pGW(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qGW(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rGW(n, mu, sigma, nu)

hGW(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dGW} gives the density, \code{pGW} gives the distribution 
function, \code{qGW} gives the quantile function, \code{rGW}
generates random deviates and \code{hGW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Gamma Weibull distribution 
with parameters \code{mu}, \code{sigma} and \code{nu}
}
\details{
The Gamma Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x) = (\frac{\sigma \mu ^{\nu}}{\tau (\nu)} x^{\nu \sigma - 1}) \exp({-\mu x^{\sigma} }),}

for x > 0.
}
\examples{
 
## The probability density function
curve(dGW(x, mu=1, sigma=2, nu=1), from=0.0001, to=4,
      col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pGW(x, mu=1, sigma=2, nu=1),
      from=0.0001, to=4, col="red", las=1, ylab="F(x)")
curve(pGW(x, mu=1, sigma=2, nu=1, lower.tail=FALSE),
      from=0.0001, to=4, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qGW(p, mu=1, sigma=2, nu=1), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pGW(x, mu=1, sigma=2, nu=1), 
      from=0.1, add=TRUE, col="red")

## The random function
hist(rGW(n=10000, mu=1, sigma=2, nu=1), freq=FALSE,
     xlab="x", las=1, main="")
curve(dGW(x, mu=1, sigma=2, nu=1),
      from=0.01, to=4, add=TRUE, col="red")

## The Hazard function
curve(hGW(x, mu=1, sigma=2, nu=1), from=1, to=6,
     col="red", ylab="Hazard function", las=1)

}
