% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMCWEI.R
\name{dMCWEI}
\alias{dMCWEI}
\alias{pMCWEI}
\alias{qMCWEI}
\alias{rMCWEI}
\alias{hMCWEI}
\title{The Modified Cosine-Weibull distribution (MCWEI)}
\usage{
dMCWEI(x, mu = 2.2, sigma = 1.2, nu = 0.5, log = TRUE)

pMCWEI(q, mu = 2.2, sigma = 1.2, nu = 0.5, lower.tail = TRUE, log.p = FALSE)

qMCWEI(p, mu = 2.2, sigma = 1.2, nu = 0.5, lower.tail = TRUE, log.p = FALSE)

rMCWEI(n, mu = 2.2, sigma = 1.2, nu = 0.5)

hMCWEI(x, mu = 2.2, sigma = 1.2, nu = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter representing the shape parameter \eqn{\phi} (\code{mu > 0}).}

\item{sigma}{parameter representing the scale parameter \eqn{\tau} (\code{sigma > 0}).}

\item{nu}{parameter representing the additional modified cosine parameter \eqn{\sigma} (\code{nu != 0}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMCWEI} gives the density, \code{pMCWEI} gives the cumulative distribution 
function, \code{qMCWEI} gives the quantile function, \code{rMCWEI}
generates random deviates and \code{hMCWEI} gives the hazard function.
}
\description{
Density function, cumulative distribution function, quantile function, 
random generation and hazard function for the 
Modified Cosine-Weibull distribution with 
parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The Modified Cosine-Weibull with parameters \code{mu}, \code{sigma} and \code{nu}
has density given by

\eqn{  f(x | \mu, \sigma ,\nu) = \frac{\pi \mu \sigma \nu x^{\mu -1} e^{-\sigma x^{\mu}} \cos(\frac{\pi}{2} e^{-\sigma x^{\mu}})\sin(\frac{\pi}{2} e^{-\sigma x^{\mu}})}{e^{\nu} -1} \cdot e^{\nu \left( 1- \cos^2(\frac{\pi}{2} e^{-\sigma x^{\mu}}) \right) }    }

for \eqn{x\ge 0}, \eqn{\mu>0}, \eqn{\sigma>0} and \eqn{\nu \neq 0}.
}
\examples{
# Example for MCWEI distribution. You can look for some of these figures in the original paper.

# Example 1: Plotting the density function (dMCWEI) for different values.
mu_ <- c(2.8, 4.4)
sigma_ <- c( 0.1, 0.01)
nu_ <- c( 0.1, 0.01)
X_ <- seq(0.0001, 5, length=350)

curve(dMCWEI(x, log = FALSE), type = "l", col= "royalblue", lwd=2.5,
     from = 0.001, to = 5, ylim=c(0,3), xlim = c(0,4), xlab = "x", ylab = "f(x)")
title("Density Function")

colors <- c("red", "green")

for(k in seq(1,2)) {
  curve(
    dMCWEI(x, mu=mu_[k], sigma=sigma_[k], nu = nu_[k], log = FALSE ),
    type = "l", col = colors[k], lwd=2, lty = 3, 
    from = 0.001, to=5, add = TRUE)
}

legend("topright",
       legend = c(
         expression(mu==2.2 ~ sigma==1.2 ~ nu==0.50),
         expression(mu==2.8 ~ sigma==0.1 ~ nu==0.10),
         expression(mu==4.4 ~ sigma==0.01 ~ nu==0.01)
       ),
       lwd=2,
       col=c("royalblue",colors),
       lty = c(1,3,3),
       bty="n")

# Example 2: Plotting the Cumulative Distribution function (pMCWEI) for differente values.

parameters <- data.frame(
  mu=c(0.5, 1.5, 2.5),
  sigma=c(1,0.5,1.2),
  nu=c(0.8,1,1.5))

colors <- c("royalblue", "red", "green")
curve(pMCWEI(x, mu=parameters$mu[1], 
             sigma = parameters$sigma[1], 
             nu = parameters$nu[1]),
      lwd=2.5, col=colors[1], 
      from = 0.001, to=5, 
      xlab = "x", ylab = "F(x)")
title("Cumulative Probability")

for(k in seq(2, 3)) {
curve(pMCWEI(x, 
             mu=parameters$mu[k], 
             sigma = parameters$sigma[k], 
             nu = parameters$nu[k]),
      lwd=2, col=colors[k], add = TRUE, lty=3)
  
}

legend(
  "bottomright",
  legend = c(
    expression(mu==0.5 ~" "  ~sigma==1 ~ " " ~ nu==0.8),
    expression(mu==1.5 ~" " ~  sigma==0.5 ~  " " ~ nu==1.0),
    expression(mu==2.5 ~" "~ sigma==1.2 ~ " "~nu==1.5)
  ),
  lwd = 2, lty = c(1,3,3), bty="n")


# Example 3: Plotting the Quantile function for different values

pars <- c(0.9, 0.1, -0.9)

P <- seq(0.0001, 0.9999, length=300)

curve(qMCWEI(x, mu = pars[1], sigma=pars[2], nu=pars[3]),
      col = "royalblue", lwd=3, 
      from = 0.0001, to=0.9999, xlab = "p", ylab = "Q(p)")
title("Quantile function")
legend("topleft",
  legend = c(
    expression(mu==0.5 ~" "  ~sigma==1 ~ " " ~ nu==0.8)),
  col = "royalblue",
  lty=1,
  lwd = 2)

# Example 4: Generating a Random Sample for the distribution.

set.seed(5)
hist(rMCWEI(200), freq=FALSE, col = "orange", 
     xlab = "x", ylab = "y", 
     main = "Theoretical values V.S Experimental values")
curve(dMCWEI(x, log = FALSE), add = TRUE, 
      from = 0.0001, to=5, col="royalblue", lwd=3)

legend("topright",
       legend = c(expression(f(x))),
       lwd = 2, lty = 1, bty="n", col="royalblue")

}
\references{
Rui Su, Najla M. Aloraini, Alia A. Alkhathami, Huda M. Alshanbari. 
On A new statistical distribution: Its empirical exploration using the reliability
and lifespan data in fashion industry. 
Alexandria Engineering Journal.
}
\seealso{
\link{BS}.
}
\author{
Juan Andrés Henao Arias, \email{juhenaoar@unal.edu.co}
}
