% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMuth.R
\name{dMuth}
\alias{dMuth}
\alias{pMuth}
\alias{qMuth}
\alias{rMuth}
\alias{hMuth}
\title{The Muth distribution}
\usage{
dMuth(x, mu, log = FALSE)

pMuth(q, mu, lower.tail = TRUE, log.p = FALSE)

qMuth(p, mu, lower.tail = TRUE, log.p = FALSE)

rMuth(n, mu, log.p = FALSE)

hMuth(x, mu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMuth} gives the density, \code{pMuth} gives the distribution 
function, \code{qMuth} gives the quantile function, \code{rMuth}
generates random deviates and \code{hMuth} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Power Lindley distribution 
with parameter \code{mu}.
}
\details{
The Muth distribution with parameter \code{mu} 
has density given by

\eqn{f(x)= ({exp(\mu x) - \mu})exp[{\mu x - \frac{1}{\mu} ({exp(\mu x) - 1})}]}

for \eqn{x > 0}, \eqn{0< \mu <=1}.
}
\examples{
 
## The probability density function
curve(dMuth(x, mu=1), from=0.1, to=5,
      ylim=c(0, 1), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pMuth(x, mu=1),
      from=0.1, to=2,  col="red", las=1, ylab="F(x)")
curve(pMuth(x,  mu=1, lower.tail=FALSE),
      from=0.1, to=2, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qMuth(p, mu=1), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pMuth(x, mu=1), from=0.01, add=TRUE, col="red")

## The random function
hist(rMuth(n=10000, mu=1), freq=FALSE,
     xlab="x", las=1, main="")
curve(dMuth(x, mu=1), from=0.1, to=8, add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1, 1))
curve(hMuth(x, mu=1), from=0.1, to=3.5,
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{abdullah2018}{RelDists}
}
\author{
William Aguilar
}
