% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOmega.R
\name{dOmega}
\alias{dOmega}
\alias{pOmega}
\alias{qOmega}
\alias{rOmega}
\alias{hOmega}
\title{The Omega distribution}
\usage{
dOmega(x, mu, sigma, nu, log = FALSE)

pOmega(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qOmega(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rOmega(n, mu, sigma, nu)

hOmega(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scale parameter.}

\item{sigma}{shape parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dOme} gives the density, \code{pOmega} gives the distribution
function, \code{qOmega} gives the quantile function, \code{rOmega}
generates random deviates and \code{hOmega} gives the hazard function.

\code{dOmega} gives the density, \code{pOmega} gives the distribution 
function, \code{qOmega} gives the quantile function, \code{rOmega}
generates random deviates and \code{hOmega} gives the hazard function.
}
\description{
Density, distribution function, quantile function,
random generation and hazard function for the Omega distribution
with parameters \code{mu}, \code{sigma}, \code{nu}.
}
\details{
The Omega Distribution with parameters \code{mu},
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\mu \sigma x^{\sigma - 1} \frac{\nu^{2\sigma}}{\nu^{2\sigma} - x^{2\sigma}} \omega_{\nu}^{(-\mu, \sigma)}(x),}

where

\eqn{\omega_{\nu}^{(-\mu, \sigma)}(x)=\left(\frac{\nu^{\sigma}+x^{\sigma}}{\nu^{\sigma}-x^{\sigma}}\right)^{\frac{-\mu \nu^{\sigma}}{2}},}

for \eqn{x > 0}, \eqn{\mu > 0}, \eqn{\sigma > 1} and \eqn{\nu > 0}.
}
\examples{
## The probability density function
curve(dOmega(x, mu=0.4, sigma=1.5, nu=5), from=0, to=5,
      ylim=c(0, 0.5), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pOmega(x, mu=0.4, sigma=1.5, nu=5),
      from=0, to=2,  col="red", las=1, ylab="F(x)")
curve(pOmega(x, mu=0.4, sigma=1.5, nu=8, lower.tail=FALSE),
      from=0, to=10, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qOmega(p, mu=0.3, sigma=1.5, nu=5), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pOmega(x, mu=0.3, sigma=1.5, nu=5), from=0, add=TRUE, col="red")

## The random function
hist(rOmega(n=10000, mu=0.4, sigma=1.5, nu=5), freq=FALSE,
     xlab="x", las=1, main="")
curve(dOmega(x, mu=0.4, sigma=1.5, nu=5), from=0, add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1,1))
curve(hOmega(x, mu=0.4, sigma=1.5, nu=50), from=0, to=5, ylim=c(0, 2),
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{dombi2019omega}{RelDists}
}
\author{
Edwin Caicedo
}
