% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCNB.R
\name{dPCNB}
\alias{dPCNB}
\alias{pPCNB}
\alias{qPCNB}
\alias{rPCNB}
\alias{hPCNB}
\title{The Power-Cauchy Negative-Binomial distribution}
\usage{
dPCNB(x, mu, sigma, nu, tau, log = FALSE)

pPCNB(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qPCNB(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rPCNB(n, mu, sigma, nu, tau)

hPCNB(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{tau}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dPCNB} gives the density, \code{pPCNB} gives the distribution 
function, \code{qPCNB} gives the quantile function, \code{rPCNB}
generates random deviates and \code{hPCNB} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Power-Cauchy Negative-Binomial distribution
with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
Power-Cauchy Negative-Binomial distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x) = \frac{2\pi^{-1}\nu\tau^{\nu}(\mu/\sigma)(x/\sigma)^{\mu-1}\left[1+(x/\sigma)^{2\mu}\right]^{-1}
\left[2\pi^{-1}\tan^{-1}(x/\sigma)^{\mu}\right]^{\nu-1}}
{\left[1-2\pi^{-1}(1-\tau)\tan^{-1}(x/\sigma)^{\mu}\right]^{\nu+1}},}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\nu> 0}, \eqn{ 0 <\nu< 1}
}
\examples{
 
## The probability density function
curve(dPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8), from=0.001, 
      to=3, ylim=c(0, 1), col="red", las=1, ylab="f(x)")

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8),
      from=0.01, to=2.5, col="red", las=1, ylab="F(x)")
curve(pPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8, lower.tail=FALSE),
      from=0.01, to=2.5, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qPCNB(p, mu=3, sigma=1, nu=0.9, tau=0.8), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8), 
      from=0.00001, add=TRUE, col="red")

## The random function
hist(rPCNB(n=10000, mu=3, sigma=1, nu=0.9, tau=0.8), freq=FALSE,
     xlab="x", las=1, main="")
curve(dPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8),
      from=0.0001, to=10, add=TRUE, col="red")

## The Hazard function
curve(hPCNB(x, mu=3, sigma=1, nu=0.9, tau=0.8), from=0.001, to=20,
      col="red", ylab="Hazard function", las=1)


}
\references{
\insertRef{muhammad2018}{RelDists}
}
\author{
Santiago Uribe
}
