% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dWGE.R
\name{dWGE}
\alias{dWGE}
\alias{pWGE}
\alias{qWGE}
\alias{rWGE}
\alias{hWGE}
\title{The Weighted Gamma-Exponential distribution}
\usage{
dWGE(x, mu, sigma, nu, log = FALSE)

pWGE(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qWGE(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rWGE(n, mu, sigma, nu)

hWGE(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{scale parameter.}

\item{sigma, nu}{shape parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dWGE} gives the density, \code{pWGE} gives the distribution 
function, \code{qWGE} gives the quantile function, \code{rdWGE}
generates random deviates and \code{hWGE} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Weighted Gamma-Exponential 
distribution with parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The Weighted Gamma-Exponential distribution with parameters \code{mu},
\code{sigma} and \code{nu} has density given by

\eqn{f(x)= \mu^\nu x^(\nu-1) \exp(-\mu x)(1 - \exp(-\sigma\mu x)) / \Gamma(\nu)(1 - (1 + \alpha)^(-\nu) ,}

for \eqn{x > 0}, \eqn{\mu > 0}, \eqn{\sigma > 0} and \eqn{\nu >0}.
}
\examples{
 
## The probability density function 
par(mfrow = c(1,2))
curve(dWGE(x, mu=1, sigma=1, nu=1), from=0, to=8, ylim=c(0, 0.6), 
      col="red", las=1, ylab="The probability density function")
curve(dWGE(x, mu=1, sigma=1, nu=0.5), from=0, to=8, ylim=c(0, 0.8), 
      col="red", las=1, ylab="The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pWGE(x, mu=1, sigma=1, nu=1), from=0, to=10, ylim=c(0, 1), 
      col="red", las = 1, ylab = "The cumulative distribution function")
curve(pWGE(x, mu=1, sigma=1, nu=1, lower.tail=FALSE), from=0, to=10,  
      ylim=c(0, 1), col="red", las=1, ylab="The Reliability function")

## The quantile function
p <- seq(from=0, to=0.998, length.out=100)
plot(x=qWGE(p, mu=1, sigma=1, nu=1), y=p, xlab="Quantile", 
    las=1, ylab="Probability")
curve(pWGE(x, mu=1, sigma=1, nu=1), from=0, add=TRUE, col="red")

## The random function
hist(rWGE(n=100, mu=1, sigma=1, nu=1), freq=FALSE, 
     ylim = c(0,0.6),xlab="x", las=1, main="")
curve(dWGE(x, mu=1, sigma=1, nu=1), from=0, ylim=c(0, 0.6), 
      add=TRUE, col="red")

## The Hazard function
par(mfrow=c(1,2))
curve(hWGE(x, mu=1, sigma=1, nu=1), from=0, to=8, ylim=c(0, 1.3), 
      col="red", ylab="The hazard function", las=1)
curve(hWGE(x, mu=1, sigma=1, nu=0.5), from=0.1, to=8, 
      col="red", ylab="The hazard function", las=1)

}
\references{
\insertRef{Mahdavi2015}{RelDists}
}
\author{
Jaime Mosquera Gutiérrez, \email{jmosquerag@unal.edu.co}
}
