% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EC.R
\name{EC}
\alias{EC}
\title{The Exponentiated Chen family}
\usage{
EC(mu.link = "log", sigma.link = "log", nu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}
}
\description{
The Exponentiated Chen family
}
\details{
The Exponentiated Chen distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\mu \sigma \nu x^{\sigma - 1} \exp({x ^\sigma}) \exp\{{\nu (1 -\exp({x^\sigma})}\} 
\{1 - \exp({\nu (1 -\exp({x^\sigma}))})\}^{\mu-1},}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\nu> 0}.
}
\examples{
# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rEC(n=500, mu=35, sigma=0.25, nu=2)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family='EC',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma and nu
# using the inverse link function
exp(coef(mod, what='mu'))
exp(coef(mod, what='sigma'))
exp(coef(mod, what='nu'))

# Example 2
# Generating random values under some model
n <- 2000
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(0.044 + 3 * x1)
sigma <- exp(0.397 - 2 * x2)
nu <- 2
x <- rEC(n=n, mu, sigma, nu)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, family=EC,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
coef(mod, what="sigma")
exp(coef(mod, what="nu"))


}
\references{
\insertRef{sanku2017}{RelDists}
}
\seealso{
\link{dEC}
}
