% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEG.R
\name{EEG}
\alias{EEG}
\title{The Extended Exponential Geometric family}
\usage{
EEG(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}
}
\description{
The Extended Exponential Geometric family
}
\details{
The Extended Exponential Geometric distribution with parameters \code{mu} 
and \code{sigma} has density given by

\eqn{f(x)= \mu \sigma \exp(-\mu x)(1 - (1 - \sigma)\exp(-\mu x))^{-2},}

for \eqn{x > 0}, \eqn{\mu > 0} and \eqn{\sigma > 0}.
}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{adamidis2005extension}{RelDists}
}
\seealso{
\link{dEEG}
}
\author{
Johan David Marin Benjumea, \email{johand.marin@udea.edu.co}
}
