% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equipment-data.R
\docType{data}
\name{equipment}
\alias{equipment}
\title{Electronic equipment}
\format{
A vector with 18 observations.
}
\usage{
data(equipment)
}
\description{
Time to failure in hours of 18 units of the same electronic device.
}
\examples{
data(equipment)
hist(equipment, main="")
}
\keyword{datasets}
