% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJ2.R
\name{estim_mu_sigma_CJ2}
\alias{estim_mu_sigma_CJ2}
\title{initValuesCJ2}
\usage{
estim_mu_sigma_CJ2(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\value{
A two-length numeric vector with initial estimates for \eqn{mu} and \eqn{sigma}
parameters from CJ2 distribution (see \code{\link{dCJ2}}).
}
\description{
This function generates initial values for CJ2 distribution.
}
\keyword{internal}
