% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWE.R
\name{estim_mu_sigma_FWE}
\alias{estim_mu_sigma_FWE}
\title{estim_mu_sigma_FWE}
\usage{
estim_mu_sigma_FWE(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for FWE distribution.
}
\examples{
y <- rFWE(n=100, mu=0.75, sigma=1.3)
estim_mu_sigma_FWE(y=y)
}
