% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL2.R
\name{estim_mu_sigma_GL2}
\alias{estim_mu_sigma_GL2}
\title{estim_mu_sigma_GL2}
\usage{
estim_mu_sigma_GL2(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for the GL2 distribution
}
\examples{
y <- rGL2(n = 100, mu = 3, sigma = 1.2)
estim_mu_sigma_GL2(y = y)
}
