% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLIN.R
\name{estim_mu_sigma_GLIN}
\alias{estim_mu_sigma_GLIN}
\title{estim_mu_sigma_GLIN}
\usage{
estim_mu_sigma_GLIN(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for the GLIN distribution
}
\examples{
y <- rGLIN(n = 100, mu = 3, sigma = 1.2)
estim_mu_sigma_GLIN(y = y)
}
