% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOEW.R
\name{estim_mu_sigma_MOEW}
\alias{estim_mu_sigma_MOEW}
\title{Initial values for MOEW}
\usage{
estim_mu_sigma_MOEW(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the MLE estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
