% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEE.R
\name{estim_mu_sigma_NEE}
\alias{estim_mu_sigma_NEE}
\title{Initial values for NEE distribution}
\usage{
estim_mu_sigma_NEE(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with the MLE estimations.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
