% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{exwstpt}
\alias{exwstpt}
\title{Auxiliar function for the Ex-Wald distribution}
\usage{
exwstpt(x, p = 0.5)
}
\arguments{
\item{x}{a value for x.}

\item{p}{a value for p.}
}
\value{
returns a vector with starting values.
}
\description{
This function generates start values.
}
\keyword{internal}
