% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist_func_family.R
\name{fitDistPlus}
\alias{fitDistPlus}
\title{Fitting Different Parametric  \code{gamlss.family} Distributions (\code{RelDists} inclusive).}
\usage{
fitDistPlus(y, k = 2, type, try.gamlss = FALSE, extra = NULL,
  data = NULL, trace = FALSE, ...)
}
\arguments{
\item{y}{the data vector}

\item{k}{a GAMLSS fitted model}

\item{type}{the type of distribution to be tried see details}

\item{try.gamlss}{this applies to functions fitDist() and fitDistPred(). It allows if gamlssML() fail to fit the model to try gamlss instead. This will slow up things for big data.}

\item{extra}{whether extra distributions should be tried, which are not in the type list}

\item{data}{the data frame where y can be found, only for functions fitDist() and fitDistPred()}

\item{trace}{whether to print during fitting.}

\item{...}{for extra arguments to be passed to gamlssML() to gamlss()}
}
\description{
This family of functions are a clones of \code{\link[gamlss]{fitDist}}, 
\code{\link{fitDistPred}}, \code{\link{chooseDist}} and
\code{\link{chooseDistPred}} of \code{\link[gamlss]{gamlss}} package. Type in
console \code{?gamlss::fitDist} for further information about the usage of this functions.
}
\details{
The original \code{\link[gamlss]{fitDist}} has seven different family types; \code{realAll},
\code{realline, realplus, real0to1, counts} and \code{binom}. This implementation has an
additional function:

\itemize{
  \item \code{reldists}: distributions defined in the positive real line, implemented in
  \code{RelDists} package: "AddW", "BW", "EW", "ExW", "FWE", "GammaW",  "GMW", 
  "IW", "KW", "LW", "MW", "PL", "RW", "SZMW", "WG", "WGEE", "WP".
}
}
\seealso{
\code{\link[gamlss]{fitDist}}
}
