% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{fitexw}
\alias{fitexw}
\title{Auxiliar function for the Ex-Wald distribution}
\usage{
fitexw(rt, p = 0.5, start = exwstpt(rt, p), scaleit = TRUE)
}
\arguments{
\item{rt}{a vector with the random sample.}

\item{p}{a value for p.}

\item{start}{an optional start vector.}

\item{scaleit}{logical value to scale.}
}
\value{
returns a vector with cumulative probabilities.
}
\description{
This function generates starting values.
}
\keyword{internal}
