% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAo.R
\name{GAo}
\alias{GAo}
\title{The gamma family in its original parameterization}
\usage{
GAo(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma.}
}
\value{
Returns a gamlss.family object which can be used to fit a GAo distribution in the \code{gamlss()} function.
}
\description{
The function \code{GAo()} defines The gamma, 
a two parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting 
using the function \code{gamlss()}.
}
\details{
The gamma original with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{x^{\mu-1}e^{-x/\sigma}}{\sigma^\mu \Gamma(\mu)}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. 
The parameter \eqn{\mu} is the shape parameter and \eqn{\sigma} is 
the scale parameter.   
In this parameterization \eqn{\mu} is the median of \eqn{X}, 
\eqn{E(X)=\mu \sigma} and 
\eqn{Var(X)=\mu \sigma^2}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rGAo(n=100, mu=0.75, sigma=1.3)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=GAo)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y as GAo
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.45 - 3 * x1)
  sigma <- exp(2 - 1.5 * x2)
  y <- rGAo(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=500)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=GAo, data=dat)

summary(mod2)
}
\references{
Abramowitz M, Stegun IA (1972). Handbook of Mathematical Functions 
with Formulas, Graphs, and Mathematical Tables. Dover Publications, 
New York. ISBN 0486612724. Chapter 6: Gamma and Related Functions.
}
\seealso{
\link{dGAo}
}
