% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initValuesOW_TTT.R
\name{initValuesOW_TTT}
\alias{initValuesOW_TTT}
\title{Initial values and search region for Odd Weibull distribution}
\usage{
initValuesOW_TTT(
  formula,
  data = NULL,
  local_reg = NULL,
  interpolation = interp.options(),
  ...
)
}
\arguments{
\item{formula}{an object of class \code{\link{formula}} with the response on
the left of an operator \code{~}. The right side must be 
\code{1}.}

\item{data}{an optional data frame containing the response variables. If 
data is not specified, the variables are taken from the 
environment from which \code{initValuesOW_TTT} is called.}

\item{local_reg}{a list of control parameters for LOESS. See 
\code{\link{loess.options}}.}

\item{interpolation}{a list of control parameters for interpolation function. See 
\code{\link{interp.options}}.}

\item{...}{further arguments passed to 
\code{\link[EstimationTools]{TTTE_Analytical}}.}
}
\description{
This function can be used so as to get suggestions about initial values
and the search region for parameter estimation in \code{OW} distribution.
}
\details{
This function performs a non-parametric estimation of the empirical total 
time on test (TTT) plot. Then, this estimated curve can be used so as to 
get suggestions about initial values and the search region for parameters 
based on hazard shape associated to the  shape of empirical TTT plot.
}
\examples{
# Example 1
# Bathtuh hazard and its corresponding TTT plot
y1 <- rOW(n=50, mu=0.022, sigma=8, nu=0.01)
my_initial_guess <- initValuesOW_TTT(formula=y1~1)
summary(my_initial_guess)
plot(my_initial_guess, legend_options=list(pos=1.04),
     par_plot=list(mar=c(3.7,3.7,1,1.5), mgp=c(2.5,1,0)))

curve(hOW(x, mu = 0.022, sigma = 8, nu = 0.01), from = 0, 
      to = 80, ylim = c(0, 0.04), col = "red", 
      ylab = "Hazard function", las = 1)

# Example 2
# Parameters explained with a covariate
set.seed(789)
x <- c(rep(0,25), rep(1,25))
nu <- 0.1 + 1.8*x
y2 <- rOW(n=50, mu=0.05, sigma=2, nu=nu)
x <- as.factor(x)

# partitions <- list(method='quantile-based', folds=5)
# my_initial_val <- initValuesOW_TTT(y2 ~ x, partition_method=partitions)
my_initial_val <- initValuesOW_TTT(y2 ~ x)
summary(my_initial_val)
plot(my_initial_val, legend_options=list(pos=1.03),
     par_plot=list(mar=c(3.7,3.7,1,2.8), mgp=c(2.5,1,0)))
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
\concept{initValOW}
