% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPXLIN.R
\name{IPXLIN}
\alias{IPXLIN}
\title{IPXLIN}
\usage{
IPXLIN(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Esta funcion sirve para obtener la suma de dos numeros reales.
}
\examples{
mean(c(4, 6, 8, 9, 2))
}
