% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW.R
\name{KW}
\alias{KW}
\alias{dKW}
\alias{pKW}
\alias{qKW}
\alias{rKW}
\alias{hKW}
\title{The Kumaraswamy Weibull Distribution}
\usage{
KW(mu.link = "log", sigma.link = "log", nu.link = "log",
  tau.link = "log")

dKW(x, mu, sigma, nu, tau, log = FALSE)

pKW(q, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

qKW(p, mu, sigma, nu, tau, lower.tail = TRUE, log.p = FALSE)

rKW(n, mu, sigma, nu, tau)

hKW(x, mu, sigma, nu, tau)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter one.}

\item{sigma}{parameter two.}

\item{nu}{parameter three.}

\item{tau}{parameter four.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dKW} gives the density, \code{pKW} gives the distribution 
function, \code{qKW} gives the quantile function, \code{rKW}
generates random deviates and \code{hKW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the kumaraswamy weibull distribution with
parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}.
}
\details{
The kumaraswamy weibull with parameters \code{mu}, \code{sigma}, \code{nu} and \code{tau}
has density given by

f(x)=nu*tau*mu*sigma*x^(sigma-1)*exp(-mu*x^sigma)*(1-exp(-mu*x^sigma))^(nu-1)*
(1-(1-exp(-aplha*x^sigma))^nu)^(tau-1)

for x>0.
}
\examples{
## The probability density function  
curve(dKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), from = 0, to = 3, ylim = c(0, 2.5), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), from = 0, to = 3, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5, lower.tail = FALSE), from = 0, to = 3, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(0,0.99999, length.out=100)
plot(x=qKW(p, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), from = 0, add = TRUE, col = "red")

## The random function
hist(rKW(10000, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), freq = FALSE, xlab = "x", las = 1, ylim = c(0, 2.5), main = "")
curve(dKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5),  from = 0, add = TRUE, col = "red" )

## The Hazard function
curve(hKW(x, mu = 3, sigma = 0.8, nu = 2, tau = 1.5), from = 0, to = 2.1, ylim = c(0, 4), col = "red", ylab = "The Hazard function", las = 1)
}
