% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CJ2.R
\name{logLik_CJ2}
\alias{logLik_CJ2}
\title{logLik function for CJ2}
\usage{
logLik_CJ2(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for CJ2 distribution.
}
\keyword{internal}
