% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GL2.R
\name{logLik_GL2}
\alias{logLik_GL2}
\title{logLik_GL2}
\usage{
logLik_GL2(param = c(0, 0), x)
}
\arguments{
\item{param}{Numeric vector containing the values of the parameters}

\item{x}{Numeric vector containing the observations.}
}
\description{
Auxiliary function to compute the log-likelihood of the GL2 distribution.
}
\examples{
y <- rGL2(n = 100, mu = 3, sigma = 1.2)
logLik_GL2(param = c(0, 0), x = y)

}
