% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLIN.R
\name{logLik_GLIN}
\alias{logLik_GLIN}
\title{logLik_GLIN}
\usage{
logLik_GLIN(param = c(0, 0), x)
}
\arguments{
\item{param}{Numeric vector containing the values of the parameters}

\item{x}{Numeric vector containing the observations.}
}
\description{
Auxiliary function to compute the log-likelihood of the GLIN distribution.
}
\examples{
y <- rGLIN(n = 100, mu = 3, sigma = 1.2)
logLik_GLIN(param = c(0, 0), x = y)

}
