% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOEW.R
\name{logLik_MOEW}
\alias{logLik_MOEW}
\title{logLik function for MOEW}
\usage{
logLik_MOEW(logparam = c(0, 0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for MOEW distribution.
}
\keyword{internal}
