% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEE.R
\name{logLik_NEE}
\alias{logLik_NEE}
\title{logLik function for NEE}
\usage{
logLik_NEE(logparam = c(0, 0), x)
}
\arguments{
\item{logparam}{vector with parameters in log scale.}

\item{x}{vector with the response variable.}
}
\value{
returns the loglikelihood given the parameters and random sample.
}
\description{
Calculates logLik for NEE distribution.
}
\keyword{internal}
