% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LW.R
\name{LW}
\alias{LW}
\alias{dLW}
\alias{pLW}
\alias{qLW}
\alias{rLW}
\alias{hLW}
\title{The Log-Weibull Distribution}
\usage{
LW(mu.link = "identity", sigma.link = "log")

dLW(x, mu, sigma, log = FALSE)

pLW(q, mu, sigma, lower.tail = TRUE, log.p = FALSE)

qLW(p, mu, sigma, lower.tail = TRUE, log.p = FALSE)

rLW(n, mu, sigma)

hLW(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter one.}

\item{sigma}{parameter two.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dLW} gives the density, \code{pLW} gives the distribution 
function, \code{qLW} gives the quantile function, \code{rLW}
generates random deviates and \code{hLW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the log-weibull distribution with
parameters \code{mu} and \code{sigma}.
}
\details{
The log-weibull distribution with parameters \code{mu} and 
\code{sigma} has density given by

f(x)= (1/sigma)*exp((x-mu)/sigma)*exp(-exp((x-mu)/sigma))

for -Inf < x < Inf.
}
\examples{
 
## The probability density function 
curve(dLW(x, mu = 0, sigma = 1), from = -20, to = 10, ylim = c(0, 0.4), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pLW(x, mu = 0, sigma = 1), from = 0, to = 10, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pLW(x, mu = 0, sigma = 1, lower.tail = FALSE), from = 0, to = 10, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x=qLW(p, mu = 0, sigma = 1), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pLW(x, mu = 0, sigma = 1), from = -6, add = TRUE, col = "red")

## The random function
hist(rLW(10000, mu = 0, sigma = 1), freq = FALSE,  ylim = c(0, 0.4), xlab = "x", las = 1, main = "")
curve(dLW(x, mu = 0, sigma = 1),  from = -20, to = 10,  ylim = c(0, 0.4), add = TRUE, col = "red") 

## The Hazard function
curve(hLW(x, mu = 0, sigma = 1), from = -20, to = 0, ylim = c(0, 0.3), col = "red", ylab = "The hazard function", las = 1)

}
