% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice-data.R
\docType{data}
\name{mice}
\alias{mice}
\title{Mice mortality data}
\format{
A vector with 208 data points.
}
\usage{
data(mice)
}
\description{
The ages at death in weeks for male mice exposed to 240r of gamma radiation.
}
\examples{
data(mice)
hist(mice, main="", xlab="Time (weeks)", freq=FALSE)
lines(density(mice), col="blue", lwd=2)
}
\keyword{datasets}
