% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOLE.R
\name{MOLE}
\alias{MOLE}
\title{The Marshall-Olkin Logistic-Exponential family}
\usage{
MOLE(mu.link = "log", sigma.link = "log", nu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}
}
\description{
The Marshall-Olkin Logistic-Exponential family
}
\details{
The Marshall-Olkin Logistic-Exponential distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)= \frac{ \mu \sigma \nu \exp^{\nu x}[\exp^{\nu x}-1]^{-\mu-1}} {[1 + \sigma [\exp^{\nu x}-1]^{-\mu}]^2}}

for \eqn{x > 0}, \eqn{\mu> 0}, \eqn{\sigma> 0}, \eqn{\nu> 1}.
}
\examples{
# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rMOLE(n=300, mu=0.6, sigma=5.5, nu=1)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family='MOLE',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma and nu
# using the inverse link function
exp(coef(mod, what='mu'))
exp(coef(mod, what='sigma'))
exp(coef(mod, what='nu'))

# Example 2
# Generating random values under some model
n <- 2000
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(-2.010 + 3 * x1)
sigma <- exp(2.704 -2 * x2)
nu <- 1
x <- rMOLE(n=n, mu, sigma, nu)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, family=MOLE,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
coef(mod, what="sigma")
exp(coef(mod, what="nu"))

}
\references{
\insertRef{mansoor2018}{RelDists}
}
\seealso{
\link{dMOLE}
}
