% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Muth.R
\name{Muth}
\alias{Muth}
\title{The Muth family}
\usage{
Muth(mu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}
}
\description{
The Muth family
}
\details{
The Muth distribution with parameter \code{mu} has density given by

\eqn{f(x)= ({exp(\mu x) - \mu})exp[{\mu x - \frac{1}{\mu} ({exp(\mu x) - 1})}]}

for \eqn{x > 0}, \eqn{0< \mu <=1}.
}
\examples{
# Example 1
# Generating some random values with
# known mu
y <- rMuth(n=200, mu=0.5)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, family='Muth',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu
# using the inverse link function
exp(coef(mod, what='mu'))

# Example 2
# Generating random values under some model
n <- 2000
x1 <- runif(n, min=0.4, max=0.6)
mu <- exp(-2.19 + 3 * x1)
x <- rMuth(n=n, mu)

mod <- gamlss(x~x1, family=Muth,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")

}
\references{
\insertRef{abdullah2018}{RelDists}
}
\seealso{
\link{dMuth}
}
