% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWEx.R
\name{MWEx}
\alias{MWEx}
\alias{dMWEx}
\alias{pMWEx}
\alias{qMWEx}
\alias{rMWEx}
\alias{hMWEx}
\title{The Modified Weibull Extension Distribution}
\usage{
MWEx(mu.link = "log", sigma.link = "log", nu.link = "log")

dMWEx(x, mu, sigma, nu, log = FALSE)

pMWEx(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qMWEx(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rMWEx(n, mu, sigma, nu)

hMWEx(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter one.}

\item{sigma}{parameter two.}

\item{nu}{parameter three.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dMWEx} gives the density, \code{pMWEx} gives the distribution 
function, \code{qMWEx} gives the quantile function, \code{rMWEx}
generates random deviates and \code{hMWEx} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the modified weibull extension distribution with
parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The modified weibull extension distribution with parameters \code{mu}, \code{sigma}
and \code{nu} has density given by

f(x) = nu*sigma*((x/mu)^(sigma-1))*exp((x/mu)^sigma + nu*mu*(1-exp((x/mu)^sigma)))

for x > 0.
}
\examples{
 
## The probability density function  
curve(dMWEx(x, mu = 1/0.5, sigma = 3, nu = 2), from = 0, to = 2.5, ylim = c(0, 1.5), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pMWEx(x, mu = 1/0.5, sigma = 3, nu = 2), from = 0, to = 2.5, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pMWEx(x, mu = 1/0.5, sigma = 3, nu = 2,  lower.tail = FALSE), from = 0, to = 2.5, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x = qMWEx(p, mu = 1/0.5, sigma = 3, nu = 2), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pMWEx(x, mu = 1/0.5, sigma = 3, nu = 2), from = 0, add = TRUE, col = "red")

## The random function
hist(rMWEx(n = 10000, mu = 1/0.5, sigma = 3, nu = 2), freq = FALSE, ylim = c(0, 1.5), xlab = "x", las = 1, main = "")
curve(dMWEx(x, mu = 1/0.5, sigma = 3, nu = 2),  from = 0, ylim = c(0, 2.5), add = T, col = "red")

## The Hazard function
curve(hMWEx(x, mu = 1/0.5, sigma = 3, nu = 2), from = 0, to = 1.7, ylim = c(0, 12), col = "red", ylab = "The hazard function", las = 1)
}
