% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEE.R
\name{NEE}
\alias{NEE}
\title{New Exponentiated Exponential family}
\usage{
NEE(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "logit" link as the default for the sigma.}
}
\value{
Returns a gamlss.family object which can be used to fit a 
NEE distribution in the \code{gamlss()} function.
}
\description{
The function \code{NEE()} defines the New Exponentiated Exponential distribution, a two parameter
distribution, for a \code{gamlss.family} object to be used in GAMLSS fitting
using the function \code{gamlss()}.
}
\details{
The New Exponentiated Exponential distribution with parameters \code{mu} 
and \code{sigma} has density given by

\eqn{f(x | \mu, \sigma) = \log(2^\sigma) \mu \exp(-\mu x) (1-\exp(-\mu x))^{\sigma-1} 2^{(1-\exp(-\mu x))^\sigma}, }

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}.

Note: In this implementation we changed the original parameters 
\eqn{\theta} for \eqn{\mu} and \eqn{\alpha} for \eqn{\sigma},
we did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rNEE(n=500, mu=2.5, sigma=3.5)

# Fitting the model
require(gamlss)

mod1 <- gamlss(y~1, sigma.fo=~1, family=NEE,
               control=gamlss.control(n.cyc=5000, trace=TRUE))

# Extracting the fitted values for mu, sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values under some model
gendat <- function(n) {
  x1 <- runif(n, min=0, max=5)
  x2 <- runif(n, min=0, max=5)
  mu <- exp(-0.2 + 1.5 * x1)
  sigma <- exp(1 - 0.7 * x2)
  y <- rNEE(n=n, mu, sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
datos <- gendat(n=500)

mod2 <- gamlss(y~x1, sigma.fo=~x2, family=NEE, data=datos,
               control=gamlss.control(n.cyc=5000, trace=TRUE))

summary(mod2)

}
\references{
Hassan, Anwar, I. H. Dar, and M. A. Lone. "A New Class of Probability 
Distributions With An Application to Engineering Data." 
Pakistan Journal of Statistics and Operation Research 20.2 (2024): 217-231.
}
\seealso{
\link{dNEE}
}
