% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{negllexw}
\alias{negllexw}
\title{Auxiliar function to obtain minus loglik for Ex-Wald}
\usage{
negllexw(p, x)
}
\arguments{
\item{p}{a vector with the parameters.}

\item{x}{a vector with the random sample.}
}
\value{
returns the minus loglik.
}
\description{
This function calculates minus loglik.
}
\keyword{internal}
