% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.start.R
\name{param.start}
\alias{param.start}
\title{Initial values extraction for Odd Weibull distribution}
\usage{
param.start(param, initValOW)
}
\arguments{
\item{param}{a character used to specify the parameter required. It can take the
values \code{"sigma"} or \code{"nu"}.}

\item{initValOW}{an \code{initValOW} object generated with \code{\link{initValuesOW_TTT}}
function.}
}
\description{
This function can be used to extract initial values found with empirical 
time on test transform (TTT) through  \code{\link{initValuesOW_TTT}} function. 
This is used for parameter estimation in \code{OW} distribution.
}
\details{
This function just gets initial values computed with \code{\link{initValuesOW_TTT}} 
for \code{OW} family. It must be called in \code{sigma.start} and \code{nu.start} 
arguments from \code{\link[gamlss]{gamlss}} function. This function is useful only
if user want to set start values automatically with TTT plot.
See example for an illustration.
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
