% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.initVal.R
\name{plot.initValOW}
\alias{plot.initValOW}
\title{Plot of \code{initValOW} objects}
\usage{
\method{plot}{initValOW}(
  x,
  xlab = "i/n",
  ylab = expression(phi(u)),
  xlim = c(0, 1),
  ylim = c(0, 1),
  col = 1,
  lty = NULL,
  lwd = NA,
  main = "",
  curve_options = list(col = 2, lwd = 2, lty = 1),
  legend_options = list(pos = "top"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{initVal}, generated with 
\code{\link{initValuesOW_TTT}}.}

\item{xlab, ylab}{titles for x and y axes, as in \code{\link{plot}}.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits (x1, x2) of the plot.}

\item{col}{The colors for lines and points. Multiple colors can be specified. This is 
the usual color argument of \code{\link[graphics]{plot.default}}.}

\item{lty}{a vector of line types, see \code{\link{par}} for further information.}

\item{lwd}{a vector of line widths, see \code{\link{par}} for further information.}

\item{main}{a main title for the plot,}

\item{curve_options}{a list with further arguments useful for customization
of non-parametric estimate plot.}

\item{legend_options}{a list with fur further arguments useful for customization
of the legend of the plot.}

\item{...}{further arguments passed empirical TTT plot.}
}
\description{
Draws the empirical total time on test (TTT) plot and its non-parametric
(LOESS) estimated curve useful for searching initial values for \code{\link{OW}} 
family.
}
\details{
This plot complements the use of \code{\link{initValuesOW_TTT}}. It is always 
advisable to use this function in order to check the result of non-parametric estimate 
of TTT plot. See the first example in \strong{Examples} section for an illustration.

The possible arguments for \code{...} can be consulted in 
\code{\link[graphics]{plot.default}}  and \code{\link{par}}.
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
