% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{pwald}
\alias{pwald}
\title{Auxiliar pwald function for the Ex-Wald distribution}
\usage{
pwald(w, m, a, s = 0)
}
\arguments{
\item{w}{a value for w.}

\item{m}{a value for m.}

\item{a}{a value for a.}

\item{s}{a value for s.}
}
\value{
returns a vector with cumulative probabilities.
}
\description{
This function emulates the pWALD function.
}
\keyword{internal}
