% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExWALD.R
\name{rew}
\alias{rew}
\title{Auxiliar function to generate random values for Ex-Wald distribution}
\usage{
rew(x, y)
}
\arguments{
\item{x}{a value for x.}

\item{y}{a value for y.}
}
\value{
returns a vector with values.
}
\description{
This function generates another form to generate values.
}
\keyword{internal}
