% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.initValOW.R
\name{summary.initValOW}
\alias{summary.initValOW}
\title{Summary of \code{initValOW} objects}
\usage{
\method{summary}{initValOW}(object, ...)
}
\arguments{
\item{object}{an object of class \code{initVal}, generated with 
\code{\link{initValuesOW}}.}

\item{...}{extra arguments}
}
\description{
This \code{summary} method displays initial values and search regions
for \code{\link{OW}} family.
}
\author{
Jaime Mosquera Gutiérrez \email{jmosquerag@unal.edu.co}
}
