% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dExWALD.R
\name{uandv}
\alias{uandv}
\title{Auxiliar function for the Ex-Wald distribution}
\usage{
uandv(
  x,
  y,
  firstblock = 20,
  block = 0,
  tol = .Machine$double.eps^(2/3),
  maxseries = 20
)
}
\arguments{
\item{x}{a value for x.}

\item{y}{a value for y.}

\item{block}{a value.}

\item{tol}{is the tolerance.}

\item{maxseries}{maximum value for series.}
}
\value{
returns a vector with starting values.
}
\description{
This function is an auxiliar function.
}
\keyword{internal}
