% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WALD.R
\name{wald_start}
\alias{wald_start}
\title{Initial values for WALD}
\usage{
wald_start(y)
}
\arguments{
\item{y}{vector with the response variable.}
}
\value{
returns a vector with starting values.
}
\description{
This function generates initial values for the parameters.
}
\keyword{internal}
