% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS.R
\name{BS}
\alias{BS}
\title{The Birnbaum-Saunders family}
\usage{
BS(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma.}
}
\value{
Returns a gamlss.family object which can be used to fit a 
BS distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS()} defines The Birnbaum-Saunders, 
a two parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting 
using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x) = \frac{x^{-3/2}(x+\mu)}{2\sigma\sqrt{2\pi\mu}} \exp\left(\frac{-1}{2\sigma^2}(\frac{x}{\mu}+\frac{\mu}{x}-2)\right)}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this 
parameterization \eqn{\mu} is the median of \eqn{X}, 
\eqn{E(X)=\mu(1+\sigma^2/2)} and 
\eqn{Var(X)=(\mu\sigma)^2(1+5\sigma^2/4)}. The functions 
proposed here
corresponds to the functions created by Roquim et al. (2021)
with minor modifications to obtain correct log-likelihoods
and random samples.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rBS(n=100, mu=0.75, sigma=1.3)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))


# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.45 - 3 * x1)
  sigma <- exp(2 - 1.5 * x2)
  y <- rBS(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=300)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS, data=dat)

summary(mod2)

# Example 3
}
\references{
Roquim, F. V., Ramires, T. G., Nakamura, L. R., Righetto, A. J., 
Lima, R. R., & Gomes, R. A. (2021). Building flexible regression 
models: including the Birnbaum-Saunders distribution in the 
gamlss package. Semina: Ciências Exatas e Tecnológicas, 
42(2), 163-168.
}
\seealso{
\link{BS}.
}
