% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS10.R
\name{BS10}
\alias{BS10}
\title{The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P8 Based on the the first Tweedie)}
\usage{
BS10(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter (representing \eqn{\tau}).}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter (representing \eqn{\omega}).}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS10 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS10()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma} 
(where \code{mu} represents \eqn{\tau} and \code{sigma} represents \eqn{\omega}) 
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{\mu^2} \left[ \frac{2x}{\mu^2} + \frac{\mu^2}{2x} - 2 \right] \right) \frac{[2x + \mu^2]\sqrt{\sigma}}{2\mu^2 \sqrt{x^3}}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{\mu^2}{2} + \frac{\mu^4}{8\sigma}} and 
\eqn{Var(X) = \frac{\mu^6}{8\sigma} + \frac{5\mu^8}{64\sigma^2}}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
set.seed(12345)
y <- rBS10(n=100, mu=0.75, sigma=5)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS10)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS10
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(0.5 - 1.6 * x1)      # Aprox 0.75
  sigma <- exp(2.2 - 1.2 * x2)   # Aprox 5
  y <- rBS10(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS10, data=dat)

summary(mod2)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{dBS10}.
}
