% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS11.R
\name{BS11}
\alias{BS11}
\title{The Birnbaum-Saunders family - Santos-Neto et al. (2012) (P9 Based on the second Tweedie)}
\usage{
BS11(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter (representing \eqn{\beta}).}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter (representing \eqn{\omega}).}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS11 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS11()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma} 
(where \code{mu} represents \eqn{\beta} and \code{sigma} represents \eqn{\omega}) 
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{\sigma}{2\mu} \left[ \frac{x}{\mu} + \frac{\mu}{x} - 2 \right] \right) \frac{[x + \mu]\sqrt{\sigma}}{2\mu\sqrt{x^3}}}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \mu + \frac{\mu^2}{2\sigma}} and 
\eqn{Var(X) = \frac{\mu^3}{\sigma} + \frac{5\mu^4}{4\sigma^2}}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
set.seed(1234)
y <- rBS11(n=100, mu=1, sigma=12)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS11)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS11
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(0.5 - 1 * x1)      # Aprox 1
  sigma <- exp(1.9 + 1.2 * x2)   # Aprox 12
  y <- rBS11(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(123)
dat <- gendat(n=200)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS11, data=dat)

summary(mod2)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{dBS11}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
