% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS4.R
\name{BS4}
\alias{BS4}
\title{The Birnbaum-Saunders family - Second parameterization (Ahmed et al., 2008)}
\usage{
BS4(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma parameter.}
}
\value{
Returns a \code{gamlss.family} object which can be used to fit a 
BS4 distribution in the \code{gamlss()} function.
}
\description{
The function \code{BS4()} defines the Birnbaum-Saunders distribution, 
a two-parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting using the function \code{gamlss()}.
}
\details{
The Birnbaum-Saunders distribution with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{2\sqrt{2\pi}} \left[ \frac{\sigma}{x\sqrt{x}} + \frac{\mu}{\sqrt{x}} \right] \exp\left( -\frac{1}{2} \left[ \frac{\sigma}{\sqrt{x}} - \mu\sqrt{x} \right]^2 \right)}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{\sigma \mu + 1/2}{\mu^2}} and 
\eqn{Var(X) = \frac{\sigma \mu + 5/4}{\mu^4}}.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rBS4(n=50, mu=2, sigma=0.2)

# Fitting the model
require(gamlss)
mod1 <- gamlss(y~1, sigma.fo=~1, family=BS4)

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod1, what="mu"))
exp(coef(mod1, what="sigma"))

# Example 2
# Generating random values for a regression model

# A function to simulate a data set with Y ~ BS4
\dontrun{
gendat <- function(n) {
  x1 <- runif(n)
  x2 <- runif(n)
  mu <- exp(1.45 - 3 * x1)
  sigma <- exp(2 - 1.5 * x2)
  y <- rBS4(n=n, mu=mu, sigma=sigma)
  data.frame(y=y, x1=x1, x2=x2)
}

set.seed(1234)
dat <- gendat(n=100)

mod2 <- gamlss(y~x1, sigma.fo=~x2, 
               family=BS4, data=dat,
               control=gamlss.control(n.cyc=100))

summary(mod2)
}

# Example 3
# The response variable is the ratio between the average
# rent per acre planted with alfalfa and the corresponding 
# average rent for other agricultural uses. The density of
# dairy cows (X2, number per square mile) is the explanatory variable. 
library(alr4)
data("landrent")

landrent$ratio <- landrent$Y / landrent$X1

with(landrent, plot(x=X2, y=ratio))

mod3 <- gamlss(ratio~X2, sigma.fo=~X2, 
               data=landrent, family=BS4)

summary(mod3)
logLik(mod3)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.

Ahmed, S. E., et al. (2008). Inference in an applied accelerated life test model 
based on the Birnbaum-Saunders distribution. Journal of Statistical Computation 
and Simulation, 78(9), 809-823.
}
\seealso{
\link{dBS4}.
}
