% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBS5.R
\name{dBS5}
\alias{dBS5}
\alias{pBS5}
\alias{qBS5}
\alias{rBS5}
\alias{hBS5}
\title{The Birnbaum-Saunders distribution - Santos-Neto et al. (2012) (P3 Based on GLM)}
\usage{
dBS5(x, mu = 1, sigma = 0.5, log = FALSE)

pBS5(q, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

qBS5(p, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

rBS5(n, mu = 1, sigma = 0.5)

hBS5(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter (\code{mu > 0}).}

\item{sigma}{precision parameter \eqn{\delta} (\code{sigma > 0}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dBS5} gives the density, \code{pBS5} gives the distribution 
function, \code{qBS5} gives the quantile function, \code{rBS5}
generates random deviates and \code{hBS5} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the 
Birnbaum-Saunders distribution with 
parameters \code{mu} and \code{sigma}.
}
\details{
The Birnbaum-Saunders with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{\exp(\sigma/2)\sqrt{\sigma+1}}{4\sqrt{\pi\mu}x^{3/2}} \left[ x + \frac{\sigma\mu}{\sigma+1} \right] \exp\left( -\frac{\sigma}{4} \left[ \frac{x(\sigma+1)}{\sigma\mu} + \frac{\sigma\mu}{x(\sigma+1)} \right] \right)}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this 
parameterization 
\eqn{E(X) = \mu} and 
\eqn{Var(X) = \mu^2 \left[ \frac{2\sigma+5}{(\sigma+1)^2} \right]}.
}
\examples{
# Example 1
# Plotting the mass function for different parameter values
curve(dBS5(x, mu=1, sigma=2), 
      from=0.001, to=2,
      ylim=c(0, 1.5), 
      col="royalblue1", lwd=2, 
      main="Density function",
      xlab="x", ylab="f(x)")
curve(dBS5(x, mu=1, sigma=25),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("topright", legend=c("mu=1, sigma=2", 
                            "mu=1, sigma=25"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.6)

# Example 2
# Checking if the cumulative curves converge to 1
curve(pBS5(x, mu=1, sigma=2), 
      from=0.00001, to=6,
      ylim=c(0, 1), 
      col="royalblue1", lwd=2, 
      main="Cumulative Distribution Function",
      xlab="x", ylab="F(x)")
curve(pBS5(x, mu=1, sigma=25),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("bottomright", legend=c("mu=1, sigma=2", 
                               "mu=1, sigma=25"),
       col=c("royalblue1", "tomato", "seagreen"), lwd=2, cex=0.5)

# Example 3
# The quantile function
p <- seq(from=0, to=0.999, length.out=100)
plot(x=qBS5(p, mu=1, sigma=2), y=p, xlab="Quantile",
     las=1, ylab="Probability", main="Quantile function ")
curve(pBS5(x, mu=1, sigma=2), 
      from=0, add=TRUE, col="tomato", lwd=2.5)

# Example 4
# The random function
x <- rBS5(n=10000, mu=1, sigma=25)
hist(x, freq=FALSE)
curve(dBS5(x, mu=1, sigma=25),  
      add=TRUE, col="tomato", lwd=2)

# Example 5
# The Hazard function
curve(hBS5(x, mu=1, sigma=25), from=0.001, to=6,
      col="tomato", ylab="Hazard function", las=1)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{BS5}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
