% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBS7.R
\name{dBS7}
\alias{dBS7}
\alias{pBS7}
\alias{qBS7}
\alias{rBS7}
\alias{hBS7}
\title{The Birnbaum-Saunders distribution - Santos-Neto et al. (2012) (P5 Based on the variance)}
\usage{
dBS7(x, mu = 0.5, sigma = 10, log = FALSE)

pBS7(q, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

qBS7(p, mu = 1, sigma = 0.5, lower.tail = TRUE, log.p = FALSE)

rBS7(n, mu = 1, sigma = 0.5)

hBS7(x, mu, sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter representing the shape (\code{mu > 0}).}

\item{sigma}{parameter representing the variance (\code{sigma > 0}).}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dBS7} gives the density, \code{pBS7} gives the distribution 
function, \code{qBS7} gives the quantile function, \code{rBS7}
generates random deviates and \code{hBS7} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the 
Birnbaum-Saunders distribution with 
parameters \code{mu} and \code{sigma}.
}
\details{
The Birnbaum-Saunders with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x|\mu,\sigma) = \frac{1}{\sqrt{2\pi}} \exp\left( -\frac{1}{2\mu^2} \left[ \frac{\mu\sqrt{4+5\mu^2}}{2\sqrt{\sigma}x^{-1}} + \frac{2\sqrt{\sigma}\{x\mu\}^{-1}}{\sqrt{4+5\mu^2}} - 2 \right] \right) \times \left[ \frac{\{x\mu\}^{-1/2}\{4+5\mu^2\}^{1/4}}{2^{3/2}\sigma^{1/4}} + \frac{\sigma^{1/4}}{\{x\mu\}^{3/2}\sqrt{2}\{4+5\mu^2\}^{1/4}} \right]}

for \eqn{x>0}, \eqn{\mu>0} and \eqn{\sigma>0}. In this parameterization, 
\eqn{E(X) = \frac{[2+\mu^2]\sqrt{\sigma}}{\mu\sqrt{4+5\mu^2}}} and 
\eqn{Var(X) = \sigma}.
}
\examples{
# Example 1
# Plotting the mass function for different parameter values
curve(dBS7(x, mu=0.1, sigma=10), 
      from=0.001, to=40,
      ylim=c(0, 0.20),
      col="royalblue1", lwd=2, 
      main="Density function",
      xlab="x", ylab="f(x)")
curve(dBS7(x, mu=0.5, sigma=10),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("topright", legend=c("mu=0.1, sigma=10", 
                            "mu=0.5, sigma=10"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.6)

# Example 2
# Checking if the cumulative curves converge to 1
curve(pBS7(x, mu=0.1, sigma=10), 
      from=0.00001, to=50,
      ylim=c(0, 1), 
      col="royalblue1", lwd=2, 
      main="Cumulative Distribution Function",
      xlab="x", ylab="F(x)")
curve(pBS7(x, mu=0.5, sigma=10),
      col="tomato", 
      lwd=2,
      add=TRUE)
legend("bottomright", legend=c("mu=0.1, sigma=10", 
                               "mu=0.5, sigma=10"),
       col=c("royalblue1", "tomato"), lwd=2, cex=0.5)

# Example 3
# The quantile function
p <- seq(from=0, to=0.999, length.out=100)
plot(x=qBS7(p, mu=0.1, sigma=10), y=p, xlab="Quantile",
     las=1, ylab="Probability", main="Quantile function ")
curve(pBS7(x, mu=0.1, sigma=10), 
      from=0, add=TRUE, col="tomato", lwd=2.5)

# Example 4
# The random function
x <- rBS7(n=10000, mu=0.1, sigma=10)
hist(x, freq=FALSE)
curve(dBS7(x, mu=0.1, sigma=10),  
      add=TRUE, col="tomato", lwd=2)

# Example 5
# The Hazard function
curve(hBS7(x, mu=0.1, sigma=10), from=0.001, to=60,
      col="tomato", ylab="Hazard function", las=1)

}
\references{
Santos-Neto, M., Cysneiros, F. J. A., Leiva, V., & Ahmed, S. E. (2012). 
On new parameterizations of the Birnbaum-Saunders distribution. 
Pakistan Journal of Statistics, 28(1), 1-26.
}
\seealso{
\link{BS7}.
}
\author{
David Villegas Ceballos, \email{david.villegas1@udea.edu.co}
}
