% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGWF.R
\name{dGWF}
\alias{dGWF}
\alias{pGWF}
\alias{qGWF}
\alias{rGWF}
\alias{hGWF}
\title{The Generalized Weibull Distribution}
\usage{
dGWF(x, mu, sigma, nu, log = FALSE)

pGWF(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qGWF(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rGWF(n, mu, sigma, nu)

hGWF(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dGWF} gives the density, \code{pGWF} gives the distribution 
function, \code{qGWF} gives the quantile function, \code{rGWF}
generates random deviates and \code{hGWF} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Generalized Weibull distribution 
with parameters \code{mu}, \code{sigma}, and \code{nu}.
}
\details{
The Generalized Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)= \mu \sigma x^{\sigma - 1} (1 - \mu \nu x^\sigma)^{\frac{1}{\nu} - 1},}

for \eqn{x > 0}, \eqn{\mu > 0}, \eqn{\sigma > 0} and \eqn{0 < \nu < 1}.
}
\examples{

## The probability density function 
par(mfrow = c(1, 1))
curve(dGWF(x, mu = 0.01, sigma = 2.9, nu=0.2), from = 0, to = 8.5, 
      col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pGWF(x, mu = 0.01, sigma = 2.9, nu=0.2), from = 0, to = 8.5, 
      ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pGWF(x, mu = 0.01, sigma = 2.9, nu=0.2, lower.tail = FALSE), 
      from = 0, to = 6, ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.99999, length.out = 100)
plot(x = qGWF(p=p, mu = 0.01, sigma = 2.9, nu=0.2), y = p, 
     xlab = "Quantile", las = 1, ylab = "Probability")
curve(pGWF(x, mu = 0.01, sigma = 2.9, nu=0.2), from = 0, add = TRUE, 
      col = "red")

## The random function
hist(rGWF(1000, mu = 0.01, sigma = 2.9, nu=0.2), freq = FALSE, xlab = "x", 
     las = 1, ylim = c(0, 0.3), main = "")
curve(dGWF(x, mu = 0.01, sigma = 2.9, nu=0.2), from = 0, to =8, add = TRUE, 
      col = "red")

## The Hazard function
par(mfrow=c(1,1))
curve(hGWF(x, mu = 0.01, sigma = 2.9, nu=0.2), from = 0, to = 3, 
      ylim = c(0, 0.26), col = "red", ylab = "The hazard function", las = 1)

}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{mudholkar1994generalized}{RelDists}
}
\author{
Johan David Marin Benjumea, \email{johand.marin@udea.edu.co}
}
