% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dRNMW.R
\name{dRNMW}
\alias{dRNMW}
\alias{pRNMW}
\alias{qRNMW}
\alias{rRNMW}
\alias{hRNMW}
\title{The Reduced New Modified Weibull distribution}
\usage{
dRNMW(x, mu, sigma, nu, log = FALSE)

pRNMW(q, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

qRNMW(p, mu, sigma, nu, lower.tail = TRUE, log.p = FALSE)

rRNMW(n, mu, sigma, nu)

hRNMW(x, mu, sigma, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter.}

\item{sigma}{parameter.}

\item{nu}{parameter.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dRNMW} gives the density, \code{pRNMW} gives the distribution 
function, \code{qRNMW} gives the quantile function, \code{rRNMW}
generates random deviates and \code{hRNMW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the Reduced New Modified Weibull distribution
with parameters \code{mu}, \code{sigma} and \code{nu}.
}
\details{
The Reduced New Modified Weibull distribution \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x) = \frac{1}{2 \sqrt{x}} \{\mu + \sigma(1+2 \nu x)\exp(\nu x) \}exp(-\mu \sqrt{x}-\sigma \sqrt{x} \exp(\nu x)),}

for x > 0.
}
\examples{
 
## The probability density function
curve(dRNMW(x, mu=0.5, sigma=0.015, nu=2.75), from=0.001, to=5,
      col="red", ylab="f(x)", las=1, ylim=c(0, 1))

## The cumulative distribution and the Reliability function
par(mfrow=c(1, 2))
curve(pRNMW(x, mu=0.5, sigma=0.015, nu=2.751),
      from=0.01, to=2.5, col="red", las=1, ylab="F(x)")
curve(pRNMW(x, mu=0.5, sigma=0.015, nu=2.75, lower.tail=FALSE),
      from=0.01, to=2.5, col="red", las=1, ylab="S(x)")

## The quantile function
p <- seq(from=0, to=0.99999, length.out=100)
plot(x=qRNMW(p, mu=0.5, sigma=0.015, nu=2.75), y=p, xlab="Quantile",
     las=1, ylab="Probability")
curve(pRNMW(x, mu=0.5, sigma=0.015, nu=2.75),
      from=0, add=TRUE, col="red")

## The random function
hist(rRNMW(n=10000, mu=0.5, sigma=0.015, nu=2.75), freq=FALSE,
     xlab="x", ylim=c(0, 2), las=1, main="")
curve(dRNMW(x, mu=0.5, sigma=0.015, nu=2.75),
      from=0.001, to=4, add=TRUE, col="red")

## The Hazard function
curve(hRNMW(x, mu=0.5, sigma=0.015, nu=2.75), from=0.001, to=3,
      col="red", ylab="Hazard function", las=1)

}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{almalki2018}{RelDists}
}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
