% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEE.R
\name{estim_mu_sigma_NEE}
\alias{estim_mu_sigma_NEE}
\title{estim_mu_sigma_NEE}
\usage{
estim_mu_sigma_NEE(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\description{
This function generates initial values for NEE distribution.
}
\examples{
y <- rNEE(n=100, mu=4.6, sigma=0.35)
estim_mu_sigma_NEE(y=y)
}
