% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EW.R
\name{EW}
\alias{EW}
\title{The Exponentiated Weibull family}
\usage{
EW(mu.link = "log", sigma.link = "log", nu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}
}
\description{
The Exponentiated Weibull family
}
\details{
The Exponentiated Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)=\nu \mu \sigma x^{\sigma-1} \exp(-\mu x^\sigma) (1-\exp(-\mu x^\sigma))^{\nu-1},}

for x > 0.
}
