% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXL.R
\name{EXL}
\alias{EXL}
\title{The exponentiated XLindley family}
\usage{
EXL(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default 
for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default 
for the sigma.}
}
\value{
Returns a gamlss.family object which can be used to fit a 
EXL distribution in the \code{gamlss()} function.
}
\description{
The function \code{EXL()} defines The exponentiated XLindley, 
a two parameter distribution, for a \code{gamlss.family} object 
to be used in GAMLSS fitting 
using the function \code{gamlss()}.
}
\details{
The exponentiated XLindley with parameters \code{mu} and \code{sigma}
has density given by

 \deqn{f(x) = \frac{\sigma\mu^2(2+\mu + x)\exp(-\mu x)}{(1+\mu)^2}\left[1-
 \left(1+\frac{\mu x}{(1 + \mu)^2}\right) \exp(-\mu x)\right] ^ {\sigma-1} }
for \eqn{x \geq 0}, \eqn{\mu \geq 0} and \eqn{\sigma \geq 0}.

Note: In this implementation we changed the original parameters 
\eqn{\delta} for \eqn{\mu} and \eqn{\alpha} for \eqn{\sigma} we 
did it to implement this distribution within gamlss framework.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma
y <- rEXL(n=1000, mu=0.75, sigma=1.3)

# Fitting the model
require(gamlss)
mod <- gamlss(y~1, sigma.fo=~1, family="EXL")

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod, what="mu"))
exp(coef(mod, what="sigma"))


# Example 2
# Generating random values under some model
n <- 1000
x1 <- runif(n)
x2 <- runif(n)
mu <- exp(1.45 - 3 * x1)
sigma <- exp(2 - 1.5 * x2)
y <- rEXL(n=n, mu=mu, sigma=sigma)

mod <- gamlss(y~x1, sigma.fo=~x2, family=EXL)

summary(mod)
}
\references{
Alomair, A. M., Ahmed, M., Tariq, S., Ahsan-ul-Haq, M., & Talib, J. 
(2024). An exponentiated XLindley distribution with properties, 
inference and applications. Heliyon, 10(3).
}
\seealso{
\link{EXL}.
}
\author{
Manuel Gutierrez Tangarife, \email{mgutierrezta@unal.edu.co}
}
