% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWE.R
\name{FWE}
\alias{FWE}
\title{The Flexible Weibull Extension distribution}
\usage{
FWE(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}
}
\description{
The Flexible Weibull Extension distribution
}
\details{
The Flexible Weibull extension with parameters \code{mu} and \code{sigma}
has density given by

\eqn{f(x) = (\mu + \sigma/x^2) exp(\mu x - \sigma/x) exp(-exp(\mu x-\sigma/x))}

for x>0.
}
\examples{
 
plot(1:5)

}
