% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWE.R
\name{initValuesFWE}
\alias{initValuesFWE}
\title{initValuesFWE}
\usage{
initValuesFWE(y)
}
\arguments{
\item{y}{vector with the random sample}
}
\value{
A two-length numeric vector with initial estimates for \eqn{mu} and \eqn{sigma}
parameters from FWE distribution (see \code{\link{dFWE}}).
}
\description{
This function generates initial values for FWE distribution.
}
\keyword{internal}
