% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KumIW.R
\name{KumIW}
\alias{KumIW}
\title{The Kumaraswamy Inverse Weibull family}
\usage{
KumIW(mu.link = "log", sigma.link = "log", nu.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}
}
\value{
Returns a gamlss.family object which can be used to fit a KumIW distribution in the \code{gamlss()} function.
}
\description{
The Kumaraswamy Inverse Weibull family
}
\details{
The Kumaraswamy Inverse Weibull Distribution with parameters \code{mu}, 
\code{sigma} and \code{nu} has density given by

\eqn{f(x)= \mu \sigma \nu x^{-\sigma - 1} \exp{- \mu x^{-\sigma}} (1 - \exp{- \mu x^{-\sigma}})^{\nu - 1},}

for \eqn{x > 0}, \eqn{\mu > 0}, \eqn{\sigma > 0} and \eqn{\nu > 0}. 

The KumIW distribution with \eqn{\nu=1} corresponds with the IW distribution.
}
\examples{
# Example 1
# Generating some random values with
# known mu, sigma and nu
y <- rKumIW(n=100, mu=1.5, sigma=2.3, nu=1)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, family=KumIW,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma and nu 
# using the inverse link function
exp(coef(mod, what="mu"))
exp(coef(mod, what="sigma"))
exp(coef(mod, what="nu"))

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n)
x2 <- runif(n)
mu <- exp(1 +  -1 * x1)
sigma <- exp(1 + -1 * x2)
nu <- 5
y <- rKumIW(n=n, mu=mu, sigma=sigma, nu=nu)

mod <- gamlss(y~x1, sigma.fo=~x2, nu.fo=~1, family=KumIW,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
coef(mod, what="sigma")
exp(coef(mod, what="nu"))
}
\references{
\insertRef{almalki2014modifications}{RelDists}

\insertRef{shahbaz2012kumaraswamy}{RelDists}
}
\seealso{
\link{dKumIW}
}
\author{
Freddy Hernandez, \email{fhernanb@unal.edu.co}
}
