% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW.R
\name{KW}
\alias{KW}
\title{The Kumaraswamy Weibull distribution}
\usage{
KW(mu.link = "log", sigma.link = "log", nu.link = "log",
  tau.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}

\item{nu.link}{defines the nu.link, with "log" link as the default for the nu parameter.}

\item{tau.link}{defines the tau.link, with "log" link as the default for the tau parameter.}
}
\description{
The Kumaraswamy Weibull distribution
}
\details{
The Kumaraswamy Weibull distribution with parameters \code{mu}, 
\code{sigma}, \code{nu} and \code{tau} has density given by

\eqn{f(x)=\nu \tau \mu \sigma x^{\sigma - 1} \exp^{-\mu x ^\sigma} [1 - \exp^{-\mu x ^ \sigma}]^{\nu-1}[1-(1-\exp^{-\mu x ^ \sigma})^\nu]^{\tau-1},}

for x > 0.
}
\examples{
# Example 1
# Generating some random values with
# known mu, sigma, nu and tau
y <- rKW(n=100, mu=3, sigma=0.8, nu=2.0, tau=1.5)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, nu.fo=~1, tau.fo=~1, family='KW',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu, sigma, nu and tau
# using the inverse link function
exp(coef(mod, what='mu'))
exp(coef(mod, what='sigma'))
exp(coef(mod, what='nu'))
exp(coef(mod, what='tau'))

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(2.6 + -3 * x1)
sigma <- exp(0.78 - 2 * x2)
nu <- 2
tau <- 1.5
x <- rKW(n=n, mu, sigma, nu, tau)

mod <- gamlss(x~x1, sigma.fo=~x2, nu.fo=~1, tau.fo=~1, family=KW,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
coef(mod, what="sigma")
exp(coef(mod, what="nu"))
exp(coef(mod, what="tau"))

}
\seealso{
\link{dKW}
}
