% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEE.R
\name{logLik_NEE}
\alias{logLik_NEE}
\title{logLik_NEE}
\usage{
logLik_NEE(param = c(0, 0), x)
}
\arguments{
\item{param}{vector with the values for mu and sigma}

\item{x}{vector with the data}
}
\description{
This is an auxiliar function to obtain the logLik for NEE.
}
\examples{
y <- rNEE(n=100, mu=1, sigma=1)
logLik_NEE(param=c(0, 0), x=y)
}
