% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OW.R
\name{OW}
\alias{OW}
\alias{dOW}
\alias{pOW}
\alias{qOW}
\alias{rOW}
\alias{hOW}
\title{The Odd Weibull Distribution}
\usage{
OW(mu.link = "log", sigma.link = "log", nu.link = "log")

dOW(x, mu, theta, nu, log = FALSE)

pOW(q, mu, theta, nu, lower.tail = TRUE, log.p = FALSE)

qOW(p, mu, theta, nu, lower.tail = TRUE, log.p = FALSE)

rOW(n, mu, theta, nu)

hOW(x, mu, theta, nu)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{parameter one.}

\item{theta}{parameter two.}

\item{nu}{parameter three.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are 
P[X <= x], otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dOW} gives the density, \code{pOW} gives the distribution 
function, \code{qOW} gives the quantile function, \code{rOW}
generates random deviates and \code{hOW} gives the hazard function.
}
\description{
Density, distribution function, quantile function, 
random generation and hazard function for the odd weibull distribution with
parameters \code{mu}, \code{theta} and \code{nu}.
}
\details{
The generalized power weibull with parameters \code{mu}, \code{theta}
and \code{nu} has density given by

f(x) = mu*theta*nu*x^(theta-1)*exp(mu*(x^theta))*(exp(mu*(x^theta))-1)^(nu-1)*(1+(exp(mu*(x^theta))-1)^nu)^-2

for x > 0.
}
\examples{
 
## The probability density function 
curve(dOW(x, mu = 2, theta = 3, nu = 0.2), from = 0, to = 4, ylim = c(0, 2), col = "red", las = 1, ylab = "The probability density function")

## The cumulative distribution and the Reliability function
par(mfrow = c(1, 2))
curve(pOW(x, mu = 2, theta = 3, nu = 0.2), from = 0, to = 4, ylim = c(0, 1), col = "red", las = 1, ylab = "The cumulative distribution function")
curve(pOW(x, mu = 2, theta = 3, nu = 0.2, lower.tail = FALSE), from = 0, to = 4,  ylim = c(0, 1), col = "red", las = 1, ylab = "The Reliability function")

## The quantile function
p <- seq(from = 0, to = 0.998, length.out = 100)
plot(x = qOW(p, mu = 2, theta = 3, nu = 0.2), y = p, xlab = "Quantile", las = 1, ylab = "Probability")
curve(pOW(x, mu = 2, theta = 3, nu = 0.2), from = 0, add = TRUE, col = "red")

## The random function
hist(rOW(n = 10000, mu = 2, theta = 3, nu = 0.2), freq = FALSE, ylim = c(0, 2),xlab = "x", las = 1, main = "")
curve(dOW(x, mu = 2, theta = 3, nu = 0.2),  from = 0, ylim = c(0, 2), add = T, col = "red")

## The Hazard function
curve(hOW(x, mu = 2, theta = 3, nu = 0.2), from = 0, to = 2.5, ylim = c(0, 3), col = "red", ylab = "The hazard function", las = 1)
}
