% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL.R
\name{PL}
\alias{PL}
\title{The Power Lindley family}
\usage{
PL(mu.link = "log", sigma.link = "log")
}
\arguments{
\item{mu.link}{defines the mu.link, with "log" link as the default for the mu parameter.}

\item{sigma.link}{defines the sigma.link, with "log" link as the default for the sigma.}
}
\value{
Returns a gamlss.family object which can be used to fit a PL distribution in the \code{gamlss()} function.
}
\description{
Power Lindley distribution
}
\details{
The Power Lindley Distribution with parameters \code{mu} 
and \code{sigma} has density given by

\eqn{f(x) = \frac{\mu \sigma^2}{\sigma + 1} (1 + x^\mu) x ^ {\mu - 1} \exp({-\sigma x ^\mu}),}

for x > 0.
}
\examples{
# Example 1
# Generating some random values with
# known mu and sigma 
y <- rPL(n=100, mu=1.5, sigma=0.2)

# Fitting the model
require(gamlss)

mod <- gamlss(y~1, sigma.fo=~1, family= 'PL',
              control=gamlss.control(n.cyc=5000, trace=FALSE))

# Extracting the fitted values for mu and sigma
# using the inverse link function
exp(coef(mod, 'mu'))
exp(coef(mod, 'sigma'))

# Example 2
# Generating random values under some model
n <- 200
x1 <- runif(n, min=0.4, max=0.6)
x2 <- runif(n, min=0.4, max=0.6)
mu <- exp(1.2 - 2 * x1)
sigma <- exp(0.8 - 3 * x2)
x <- rPL(n=n, mu, sigma)

mod <- gamlss(x~x1, sigma.fo=~x2, family=PL,
              control=gamlss.control(n.cyc=5000, trace=FALSE))

coef(mod, what="mu")
coef(mod, what="sigma")
}
\references{
Almalki, S. J., & Nadarajah, S. (2014). Modifications of the 
Weibull distribution: A review. Reliability Engineering & 
System Safety, 124, 32-55.

Ghitany, M. E., Al-Mutairi, D. K., Balakrishnan, N., & 
Al-Enezi, L. J. (2013). Power Lindley distribution 
and associated inference. Computational Statistics & Data 
Analysis, 64, 20-33.
}
\seealso{
\link{dPL}
}
\author{
Amylkar Urrea Montoya, \email{amylkar.urrea@udea.edu.co}
}
